naga::as_diagnostic_file_path
Trait AsDiagnosticFilePath

pub trait AsDiagnosticFilePath: Sealed {
    // Required method
    fn to_string_lossy(&self) -> Cow<'_, str>;
}

A trait that abstracts over types accepted for conversion to the most featureful path representation possible; that is:

When no_std is active, this is implemented for String, str, and Cow (i.e., Cow<'_, str>).

Otherwise, types that implement AsRef<Path> (to extract a &Path).

This type is used as the type bounds for various diagnostic rendering methods, i.e., WithSpan::emit_to_string_with_path.

Required Methods

fn to_string_lossy(&self) -> Cow<'_, str>