naga::back::dot

Struct StatementGraph

struct StatementGraph {
    nodes: Vec<&'static str>,
    flow: Vec<(usize, usize, &'static str)>,
    jumps: Vec<(usize, usize, &'static str, usize)>,
    dependencies: Vec<(usize, Handle<Expression>, &'static str)>,
    emits: Vec<(usize, Handle<Expression>)>,
    calls: Vec<(usize, Handle<Function>)>,
}

Stores information about the graph of statements

Fields
nodes: Vec<&'static str>

List of node names

flow: Vec<(usize, usize, &'static str)>

List of edges of the control flow, the items are defined as (from, to, label)

jumps: Vec<(usize, usize, &'static str, usize)>

List of implicit edges of the control flow, used for jump operations such as continue or break, the items are defined as (from, to, label, color_id)

dependencies: Vec<(usize, Handle<Expression>, &'static str)>

List of dependency relationships between a statement node and expressions

emits: Vec<(usize, Handle<Expression>)>

List of expression emitted by statement node

calls: Vec<(usize, Handle<Function>)>

List of function call by statement node

Implementations

impl StatementGraph

fn add(&mut self, block: &[Statement], targets: Targets) -> (usize, usize)

Adds a new block to the statement graph, returning the first and last node, respectively