naga::back
Function get_entry_points

fn get_entry_points(
    module: &Module,
    entry_point: Option<&(ShaderStage, String)>,
) -> Result<Range<usize>, (ShaderStage, String)>

Locate the entry point(s) to write.

If entry_point is given, and the specified entry point exists, returns a length-1 Range containing the index of that entry point. If no entry_point is given, returns the complete range of entry point indices. If entry_point is given but does not exist, returns an error.