Module naga::back::glsl

Backend for GLSL (OpenGL Shading Language).

The main structure is Writer, it maintains internal state that is used to output a Module into glsl

Supported versions

Core

330

400

410

420

430

450

ES

300

310

Structs

Features

Structure used to encode additions to GLSL that aren’t supported by all versions.

Options

Configuration used in the Writer.

PipelineOptions

A subset of options meant to be changed per pipeline.

ReflectionInfo

Reflection info for texture mappings and uniforms.

TextureMapping

Mapping between a texture and its sampler, if it exists.

Writer

Writer responsible for all code generation.

WriterFlags

Configuration flags for the Writer.

Enums

Error

A GLSL compilation error.

Version

A GLSL version.

Constants

SUPPORTED_CORE_VERSIONS

List of supported core GLSL versions.

SUPPORTED_ES_VERSIONS

List of supported es GLSL versions.

Type Definitions

BindingMap

Mapping between resources and bindings.