naga::back::hlsl::storage

Enum SubAccess

pub(super) enum SubAccess {
    BufferOffset {
        group: u32,
        offset: u32,
    },
    Offset(u32),
    Index {
        value: Handle<Expression>,
        stride: u32,
    },
}

One step in accessing a Storage global’s component or element.

Writer::temp_access_chain holds a series of these structures, describing how to compute the byte offset of a particular element or member of some global variable in the Storage address space.

Variants

BufferOffset

Fields

group: u32

offset: u32

Offset(u32)

Add the given byte offset. This is used for struct members, or known components of a vector or matrix. In all those cases, the byte offset is a compile-time constant.

Index

Scale value by stride, and add that to the current byte offset. This is used to compute the offset of an array element whose index is computed at runtime.

Fields

value: Handle<Expression>

stride: u32