naga::back::hlsl

Module writer

Structs

BindingArraySamplerInfo 🔒

Information for how to generate a binding_array<sampler> access.

EntryPointBinding 🔒

Structure contains information required for generating wrapped structure of all entry points arguments

EntryPointInterface 🔒

EpStructMember 🔒

MatrixType 🔒

Enums

Index 🔒

InterfaceKey 🔒

Io 🔒

Constants

ABS_FUNCTION 🔒

COMPARISON_SAMPLER_HEAP_VAR 🔒

DIV_FUNCTION 🔒

EXTRACT_BITS_FUNCTION 🔒

F2I32_FUNCTION 🔒

F2I64_FUNCTION 🔒

F2U32_FUNCTION 🔒

F2U64_FUNCTION 🔒

FREXP_FUNCTION 🔒

IMAGE_LOAD_EXTERNAL_FUNCTION 🔒

IMAGE_SAMPLE_BASE_CLAMP_TO_EDGE_FUNCTION 🔒

INSERT_BITS_FUNCTION 🔒

INTERNAL_PREFIX 🔒

Prefix for variables in a naga statement

LOCATION_SEMANTIC 🔒

MODF_FUNCTION 🔒

MOD_FUNCTION 🔒

NEG_FUNCTION 🔒

RAY_QUERY_TRACKER_VARIABLE_PREFIX 🔒

SAMPLER_HEAP_VAR 🔒

SAMPLE_EXTERNAL_TEXTURE_FUNCTION 🔒

SPECIAL_CBUF_TYPE 🔒

SPECIAL_CBUF_VAR 🔒

SPECIAL_FIRST_INSTANCE 🔒

SPECIAL_FIRST_VERTEX 🔒

SPECIAL_OTHER 🔒

Functions

find_matrix_in_access_chain 🔒

If base is an access chain of the form mat, mat[col], or mat[col][row], returns a tuple of the matrix, the column (vector) index (if present), and the row (scalar) index (if present).

get_global_uniform_matrix 🔒

Simpler version of get_inner_matrix_of_global_uniform that only looks at the immediate expression, rather than traversing an access chain.

get_inner_matrix_data 🔒

get_inner_matrix_of_global_uniform 🔒

Returns the matrix data if the access chain starting at base:

get_inner_matrix_of_struct_array_member 🔒

Returns the matrix data if the access chain starting at base:

is_subgroup_builtin_binding 🔒