naga::back::hlsl::writer
Function get_inner_matrix_of_struct_array_member

pub(super) fn get_inner_matrix_of_struct_array_member(
    module: &Module,
    base: Handle<Expression>,
    func_ctx: &FunctionCtx<'_>,
    direct: bool,
) -> Option<MatrixType>

Returns the matrix data if the access chain starting at base:

starts with an expression with resolved type of TypeInner::Matrix if direct = true

contains one or more expressions with resolved type of TypeInner::Array of TypeInner::Matrix

ends at an expression with resolved type of TypeInner::Struct