Module naga::back::msl

Metal Shading Language (MSL) backend

Binding model

Metal's bindings are flat per resource. Since there isn't an obvious mapping from SPIR-V's descriptor sets, we require a separate mapping provided in the options. This mapping may have one or more resource end points for each descriptor set + index pair.

Outputs

In Metal, built-in shader outputs can not be nested into structures within the output struct. If there is a structure in the outputs, and it contains any built-ins, we move them up to the root output structure that we define ourselves.

Structs

BindSource

BindTarget

InputStructIndex

Level

MemberIndex

Name

Options

OutputStructIndex

ParameterIndex

TypedGlobalVariable

Writer

Enums

Error

LocationMode

NameSource

ResolvedBinding

Constants

COMPONENTS

LOCATION_INPUT_STRUCT_NAME

OUTPUT_STRUCT_NAME

RESERVED_NAMES

Traits

AsName

Indexed

Functions

scalar_kind_string

separate

vector_size_string

write_string

Type Definitions

BindingMap