naga::back::msl::writer
Trait NameKeyExt

trait NameKeyExt {
    // Provided methods
    fn local(
        origin: FunctionOrigin,
        local_handle: Handle<LocalVariable>,
    ) -> NameKey { ... }
    fn oob_local_for_type(origin: FunctionOrigin, ty: Handle<Type>) -> NameKey { ... }
}

Provided Methods

fn local(origin: FunctionOrigin, local_handle: Handle<LocalVariable>) -> NameKey

fn oob_local_for_type(origin: FunctionOrigin, ty: Handle<Type>) -> NameKey

Return the name key for a local variable used by ReadZeroSkipWrite bounds-check policy when it needs to produce a pointer-typed result for an OOB access. These are unique per accessed type, so the second argument is a type handle. See docs for crate::back::msl.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.