naga::back

Module pipeline_constants

Enums

PipelineConstantError

Functions

adjust_block ğŸ”’

Replace every expression handle in block with its counterpart given by new_pos.

adjust_expr ğŸ”’

Replace every expression handle in expr with its counterpart given by new_pos.

adjust_stmt ğŸ”’

Replace every expression handle in stmt with its counterpart given by new_pos.

filter_emits_in_block ğŸ”’

Adjust Emit statements in block to skip needs_pre_emit expressions we have introduced.

map_value_to_literal ğŸ”’

process_function ğŸ”’

Replace all override expressions in function with fully-evaluated constants.

process_mesh_shader_overrides ğŸ”’

process_override ğŸ”’

Add a Constant to module for the override old_h.

process_overrides

Compact module and replace all overrides with constants.

process_workgroup_size_override ğŸ”’

revalidate ğŸ”’