naga::back

Struct FunctionCtx

pub struct FunctionCtx<'a> {
    pub ty: FunctionType,
    pub info: &'a FunctionInfo,
    pub expressions: &'a Arena<Expression>,
    pub named_expressions: &'a FastIndexMap<Handle<Expression>, String>,
}

Helper structure that stores data needed when writing the function

Fields
ty: FunctionType

The current function being written

info: &'a FunctionInfo

Analysis about the function

expressions: &'a Arena<Expression>

The expression arena of the current function being written

named_expressions: &'a FastIndexMap<Handle<Expression>, String>

Map of expressions that have associated variable names

Implementations

impl FunctionCtx<'_>

pub fn resolve_type<'a>( &'a self, handle: Handle<Expression>, types: &'a UniqueArena<Type>, ) -> &'a TypeInner

Helper method that resolves a type of a given expression.

pub const fn name_key(&self, local: Handle<LocalVariable>) -> NameKey

Helper method that generates a NameKey for a local in the current function

pub const fn argument_key(&self, arg: u32) -> NameKey

Helper method that generates a NameKey for a function argument.

Panics

If the function arguments are less or equal to arg

pub const fn external_texture_argument_key( &self, arg: u32, external_texture_key: ExternalTextureNameKey, ) -> NameKey

Helper method that generates a NameKey for an external texture function argument.

Panics

If the function arguments are less or equal to arg

If self.ty is not FunctionType::Function.

pub fn is_fixed_function_input( &self, expression: Handle<Expression>, module: &Module, ) -> Option<BuiltIn>

Returns true if the given expression points to a fixed-function pipeline input.