naga

Module common

Code common to the front and backends for specific languages.

Modules

diagnostic_debug 🔒

Displaying Naga IR terms in debugging output.

diagnostic_display 🔒

Displaying Naga IR terms in diagnostic output.

predeclared

Generating names for predeclared types.

wgsl

Code shared between the WGSL front and back ends.

Structs

DiagnosticDebug

A wrapper for displaying Naga IR terms in debugging output.

DiagnosticDisplay

A wrapper for displaying Naga IR terms in diagnostic output.

Traits

ForDebug

ForDebugWithTypes

Functions

vector_size_str

Helper function that returns the string corresponding to the VectorSize