naga
Module diagnostic_filter

DiagnosticFilters and supporting functionality.

Structs

ConflictingDiagnosticRuleError 🔒

An error returned by DiagnosticFilterMap::add when it encounters conflicting rules.

DiagnosticFilter

A filtering rule that modifies how diagnostics are emitted for shaders.

DiagnosticFilterMap 🔒

A map from diagnostic filters to their severity and span.

DiagnosticFilterNode

Represents a single parent-linking node in a tree of DiagnosticFilters backed by a crate::Arena.

Enums

FilterableTriggeringRule

A filterable triggering rule in a DiagnosticFilter.

Severity

A severity set on a DiagnosticFilter.

ShouldConflictOnFullDuplicate 🔒

Determines whether DiagnosticFilterMap::add should consider full duplicates a conflict.

StandardFilterableTriggeringRule

A filterable triggering rule in a DiagnosticFilter.