Module naga::front

Frontend parsers that consume binary and text shaders and load them into Modules.

Modules

glsl

Frontend for GLSL (OpenGL Shading Language).

spv

Frontend for SPIR-V (Standard Portable Intermediate Representation).

wgsl

Frontend for WGSL (WebGPU Shading Language).

Structs

Typifier

Helper processor that derives the types of all expressions.