Function naga::front::align_up

pub fn align_up(value: u32, align: u32) -> u32

Helper function used for aligning value to the next multiple of align

Notes:

align must be a power of two.

The returned value will be greater or equal to value.

Examples:

assert_eq!(0, align_up(0, 16));
assert_eq!(16, align_up(1, 16));
assert_eq!(16, align_up(16, 16));
assert_eq!(334, align_up(333, 2));
assert_eq!(384, align_up(257, 128));