Module naga::front::glsl

Frontend for GLSL (OpenGL Shading Language).

To begin, take a look at the documentation for the Parser.

Supported versions

Vulkan

440 (partial)

450

460

Structs

Error

Error returned during shader parsing.

Options

Per-shader options passed to parse.

Parser

The Parser is the central structure of the GLSL frontend.

ShaderMetadata

Additional information about the GLSL shader.

Enums

ErrorKind

Information about the cause of an error.

ExpectedToken

One of the expected tokens returned in InvalidToken.

Precision

A precision hint used in GLSL declarations.

Profile

The GLSL profile used by a shader.

TokenValue

A token passed from the lexing used in the parsing.