naga::front::glsl
Module ast

Structs

BuiltinVariations

Tracks the variations of the builtin already generated, this is needed because some builtins overloads can’t be generated unless explicitly used, since they might cause unneeded capabilities to be requested

EntryArg

FunctionCall

FunctionDeclaration

GlobalLookup

HirExpr

Overload

ParameterInfo

TypeQualifiers

VariableReference

Enums

FunctionCallKind

FunctionKind

How the function is implemented

GlobalLookupKind

HirExprKind

ParameterQualifier

Precision

A precision hint used in GLSL declarations.

Profile

The GLSL profile used by a shader.

QualifierKey

QualifierValue

StorageQualifier

StructLayout