naga::front::glsl::ast
Enum HirExprKind

pub enum HirExprKind {
    Access {
        base: Handle<HirExpr>,
        index: Handle<HirExpr>,
    },
    Select {
        base: Handle<HirExpr>,
        field: String,
    },
    Literal(Literal),
    Binary {
        left: Handle<HirExpr>,
        op: BinaryOperator,
        right: Handle<HirExpr>,
    },
    Unary {
        op: UnaryOperator,
        expr: Handle<HirExpr>,
    },
    Variable(VariableReference),
    Call(FunctionCall),
    Conditional {
        condition: Handle<HirExpr>,
        accept: Handle<HirExpr>,
        reject: Handle<HirExpr>,
    },
    Assign {
        tgt: Handle<HirExpr>,
        value: Handle<HirExpr>,
    },
    PrePostfix {
        op: BinaryOperator,
        postfix: bool,
        expr: Handle<HirExpr>,
    },
    Method {
        expr: Handle<HirExpr>,
        name: String,
        args: Vec<Handle<HirExpr>>,
    },
}

Variants

Access

Fields

base: Handle<HirExpr>

index: Handle<HirExpr>

Select

Fields

base: Handle<HirExpr>

field: String

Literal(Literal)

Binary

Fields

left: Handle<HirExpr>

op: BinaryOperator

right: Handle<HirExpr>

Unary

Fields

op: UnaryOperator

expr: Handle<HirExpr>

Variable(VariableReference)

Call(FunctionCall)

Conditional

Represents the ternary operator in glsl (:?)

Fields

condition: Handle<HirExpr>

The expression that will decide which branch to take, must evaluate to a boolean

accept: Handle<HirExpr>

The expression that will be evaluated if condition returns true

reject: Handle<HirExpr>

The expression that will be evaluated if condition returns false

Assign

Fields

tgt: Handle<HirExpr>

value: Handle<HirExpr>

PrePostfix

A prefix/postfix operator like ++

Fields

op: BinaryOperator

The operation to be performed

postfix: bool

Whether this is a postfix or a prefix

expr: Handle<HirExpr>

The target expression

Method

A method call like what.something(a, b, c)

Fields

expr: Handle<HirExpr>

expression the method call applies to (what in the example)

name: String

the method name (something in the example)

args: Vec<Handle<HirExpr>>

the arguments to the method (a, b, and c in the example)