naga::front::glsl::ast
Struct TypeQualifiers

pub struct TypeQualifiers<'a> {
    pub span: Span,
    pub storage: (StorageQualifier, Span),
    pub invariant: Option<Span>,
    pub interpolation: Option<(Interpolation, Span)>,
    pub precision: Option<(Precision, Span)>,
    pub sampling: Option<(Sampling, Span)>,
    pub storage_access: Option<(StorageAccess, Span)>,
    pub layout_qualifiers: FastHashMap<QualifierKey<'a>, (QualifierValue, Span)>,
}

Fields
span: Spanstorage: (StorageQualifier, Span)invariant: Option<Span>interpolation: Option<(Interpolation, Span)>precision: Option<(Precision, Span)>sampling: Option<(Sampling, Span)>storage_access: Option<(StorageAccess, Span)>

Memory qualifiers used in the declaration to set the storage access to be used in declarations that support it (storage images and buffers)

layout_qualifiers: FastHashMap<QualifierKey<'a>, (QualifierValue, Span)>
Implementations

impl<'a> TypeQualifiers<'a>

pub fn unused_errors(&self, errors: &mut Vec<Error>)

Appends errors with errors for all unused qualifiers

pub fn uint_layout_qualifier( &mut self, name: &'a str, errors: &mut Vec<Error>, ) -> Option<u32>

Removes the layout qualifier with name, if it exists and adds an error if it isn’t a QualifierValue::Uint

pub fn none_layout_qualifier( &mut self, name: &'a str, errors: &mut Vec<Error>, ) -> bool

Removes the layout qualifier with name, if it exists and adds an error if it isn’t a QualifierValue::None