naga::front::glsl
Module functions

Structs

ProxyWrite 🔒

Struct detailing a store operation that must happen after a function call

Enums

Conversion 🔒

Helper enum containing the type of conversion need for a call

Functions

builtin_required_variations 🔒

Helper method returning all the non standard builtin variations needed to process the function call with the passed arguments

conversion 🔒

Helper function, returns the type of conversion from source to target, if a conversion is not possible returns None.