naga::front::glsl
Struct ShaderMetadata

pub struct ShaderMetadata {
    pub version: u16,
    pub profile: Profile,
    pub stage: ShaderStage,
    pub workgroup_size: [u32; 3],
    pub early_fragment_tests: bool,
    pub extensions: FastHashSet<String>,
}

Additional information about the GLSL shader.

Stores additional information about the GLSL shader which might not be stored in the shader Module.

Fields
version: u16

The GLSL version specified in the shader through the use of the #version preprocessor directive.

profile: Profile

The GLSL profile specified in the shader through the use of the #version preprocessor directive.

stage: ShaderStage

The shader stage in the pipeline, passed to the parse method via the Options struct.

workgroup_size: [u32; 3]

The workgroup size for compute shaders, defaults to [1; 3] for compute shaders and [0; 3] for non compute shaders.

early_fragment_tests: bool

Whether or not early fragment tests where requested by the shader. Defaults to false.

extensions: FastHashSet<String>

The shader can request extensions via the #extension preprocessor directive, in the directive a behavior parameter is used to control whether the extension should be disabled, warn on usage, enabled if possible or required.

This field only stores extensions which were required or requested to be enabled if possible and they are supported.

Implementations

impl ShaderMetadata

fn reset(&mut self, stage: ShaderStage)