naga::front

Module wgsl

Frontend for WGSL (WebGPU Shading Language).

Modules

error 🔒

Formatting WGSL front end error messages.

index 🔒

lower 🔒

parse 🔒

Structs

Frontend

Options

Configuration for the whole parser run.

ParseError

Enums

EnableExtension

An enable-extension not guaranteed to be present in all environments.

ImplementedEnableExtension

A variant of EnableExtension::Implemented.

ImplementedLanguageExtension

A variant of LanguageExtension::Implemented.

LanguageExtension

A language extension recognized by Naga, but not guaranteed to be present in all environments.

UnimplementedEnableExtension

A variant of EnableExtension::Unimplemented.

UnimplementedLanguageExtension

A variant of LanguageExtension::Unimplemented.

Functions

parse_str

Type Aliases

Result 🔒