Enum naga::front::wgsl::Error

pub enum Error<'a> {

    Unexpected((Token<'a>, Range<usize>), ExpectedToken<'a>),
    UnexpectedComponents(Range<usize>),
    BadNumber(Range<usize>, NumberError),
    NegativeInt(Range<usize>),
    BadU32Constant(Range<usize>),
    BadMatrixScalarKind(Range<usize>, ScalarKind, u8),
    BadAccessor(Range<usize>),
    BadTexture(Range<usize>),
    BadTypeCast {
        span: Range<usize>,
        from_type: String,
        to_type: String,
    },
    BadTextureSampleType {
        span: Range<usize>,
        kind: ScalarKind,
        width: u8,
    },
    BadIncrDecrReferenceType(Range<usize>),
    InvalidResolve(ResolveError),
    InvalidForInitializer(Range<usize>),
    InvalidBreakIf(Range<usize>),
    InvalidGatherComponent(Range<usize>, u32),
    InvalidConstructorComponentType(Range<usize>, i32),
    InvalidIdentifierUnderscore(Range<usize>),
    ReservedIdentifierPrefix(Range<usize>),
    UnknownAddressSpace(Range<usize>),
    UnknownAttribute(Range<usize>),
    UnknownBuiltin(Range<usize>),
    UnknownAccess(Range<usize>),
    UnknownShaderStage(Range<usize>),
    UnknownIdent(Range<usize>, &'a str),
    UnknownScalarType(Range<usize>),
    UnknownType(Range<usize>),
    UnknownStorageFormat(Range<usize>),
    UnknownConservativeDepth(Range<usize>),
    SizeAttributeTooLow(Range<usize>, u32),
    AlignAttributeTooLow(Range<usize>, Alignment),
    NonPowerOfTwoAlignAttribute(Range<usize>),
    InconsistentBinding(Range<usize>),
    UnknownLocalFunction(Range<usize>),
    TypeNotConstructible(Range<usize>),
    TypeNotInferrable(Range<usize>),
    InitializationTypeMismatch(Range<usize>, String),
    MissingType(Range<usize>),
    MissingAttribute(&'static str, Range<usize>),
    InvalidAtomicPointer(Range<usize>),
    InvalidAtomicOperandType(Range<usize>),
    Pointer(&'static str, Range<usize>),
    NotPointer(Range<usize>),
    NotReference(&'static str, Range<usize>),
    ReservedKeyword(Range<usize>),
    Redefinition {
        previous: Range<usize>,
        current: Range<usize>,
    },
    Other,

}

Variants

Unexpected((Token<'a>, Range<usize>), ExpectedToken<'a>)

UnexpectedComponents(Range<usize>)

BadNumber(Range<usize>, NumberError)

NegativeInt(Range<usize>)

A negative signed integer literal where both signed and unsigned, but only non-negative literals are allowed.

BadU32Constant(Range<usize>)

BadMatrixScalarKind(Range<usize>, ScalarKind, u8)

BadAccessor(Range<usize>)

BadTexture(Range<usize>)

BadTypeCast

Fields

span: Range<usize>

from_type: String

to_type: String

BadTextureSampleType

Fields

span: Range<usize>

kind: ScalarKind

width: u8

BadIncrDecrReferenceType(Range<usize>)

InvalidResolve(ResolveError)

InvalidForInitializer(Range<usize>)

InvalidBreakIf(Range<usize>)

A break if appeared outside of a continuing block

InvalidGatherComponent(Range<usize>, u32)

InvalidConstructorComponentType(Range<usize>, i32)

InvalidIdentifierUnderscore(Range<usize>)

ReservedIdentifierPrefix(Range<usize>)

UnknownAddressSpace(Range<usize>)

UnknownAttribute(Range<usize>)

UnknownBuiltin(Range<usize>)

UnknownAccess(Range<usize>)

UnknownShaderStage(Range<usize>)

UnknownIdent(Range<usize>, &'a str)

UnknownScalarType(Range<usize>)

UnknownType(Range<usize>)

UnknownStorageFormat(Range<usize>)

UnknownConservativeDepth(Range<usize>)

SizeAttributeTooLow(Range<usize>, u32)

AlignAttributeTooLow(Range<usize>, Alignment)

NonPowerOfTwoAlignAttribute(Range<usize>)

InconsistentBinding(Range<usize>)

UnknownLocalFunction(Range<usize>)

TypeNotConstructible(Range<usize>)

TypeNotInferrable(Range<usize>)

InitializationTypeMismatch(Range<usize>, String)

MissingType(Range<usize>)

MissingAttribute(&'static str, Range<usize>)

InvalidAtomicPointer(Range<usize>)

InvalidAtomicOperandType(Range<usize>)

Pointer(&'static str, Range<usize>)

NotPointer(Range<usize>)

NotReference(&'static str, Range<usize>)

ReservedKeyword(Range<usize>)

Redefinition

Fields

previous: Range<usize>

current: Range<usize>

Other