Struct naga::front::wgsl::lower::ExpressionContext

pub struct ExpressionContext<'source, 'temp, 'out> {
    ast_expressions: &'temp Arena<Expression<'source>>,
    types: &'temp Arena<Type<'source>>,
    globals: &'temp mut FastHashMap<&'source str, LoweredGlobalDecl>,
    module: &'out mut Module,
    const_typifier: &'temp mut Typifier,
    expr_type: ExpressionContextType<'temp, 'out>,
}

State for lowering an ast::Expression to Naga IR.

ExpressionContexts come in two kinds, distinguished by the value of the expr_type field:

A Runtime context contributes naga::Expressions to a naga::Function’s runtime expression arena.

A Constant context contributes naga::Expressions to a naga::Module’s constant expression arena.

ExpressionContexts are constructed in restricted ways:

To get a Runtime ExpressionContext, call StatementContext::as_expression.

To get a Constant ExpressionContext, call GlobalContext::as_const.

You can demote a Runtime context to a Constant context by calling as_const, but there’s no way to go in the other direction, producing a runtime context from a constant one. This is because runtime expressions can refer to constant expressions, via Expression::Constant, but constant expressions can’t refer to a function’s expressions.

Not to be confused with wgsl::parse::ExpressionContext, which is for parsing the ast::Expression in the first place.

Fields
ast_expressions: &'temp Arena<Expression<'source>>types: &'temp Arena<Type<'source>>globals: &'temp mut FastHashMap<&'source str, LoweredGlobalDecl>

The map from the names of module-scope declarations to the Naga IR Handles we have built for them, owned by Lowerer::lower.

module: &'out mut Module

The IR Module we’re constructing.

const_typifier: &'temp mut Typifier

Type judgments for module::const_expressions.

expr_type: ExpressionContextType<'temp, 'out>

Whether we are lowering a constant expression or a general runtime expression, and the data needed in each case.

Implementations

impl<'source, 'temp, 'out> ExpressionContext<'source, 'temp, 'out>

fn as_const(&mut self) -> ExpressionContext<'source, '_, '_>

fn as_global(&mut self) -> GlobalContext<'source, '_, '_>

fn as_const_evaluator(&mut self) -> ConstantEvaluator<'_>

fn append_expression( &mut self, expr: Expression, span: Span ) -> Result<Handle<Expression>, Error<'source>>

fn const_access(&self, handle: Handle<Expression>) -> Option<u32>

fn get_expression_span(&self, handle: Handle<Expression>) -> Span

fn typifier(&self) -> &Typifier

fn runtime_expression_ctx( &mut self, span: Span ) -> Result<&mut RuntimeExpressionContext<'temp, 'out>, Error<'source>>

fn gather_component( &mut self, expr: Handle<Expression>, component_span: Span, gather_span: Span ) -> Result<SwizzleComponent, Error<'source>>

fn register_type( &mut self, handle: Handle<Expression> ) -> Result<Handle<Type>, Error<'source>>

Determine the type of handle, and add it to the module’s arena.

If you just need a TypeInner for handle’s type, use the resolve_inner! macro instead. This function should only be used when the type of handle needs to appear in the module’s final Arena<Type>, for example, if you’re creating a LocalVariable whose type is inferred from its initializer.

fn grow_types( &mut self, handle: Handle<Expression> ) -> Result<&mut Self, Error<'source>>

Resolve the types of all expressions up through handle.

Ensure that self.typifier has a TypeResolution for every expression in [self.function.expressions].

This does not add types to any arena. The Typifier documentation explains the steps we take to avoid filling arenas with intermediate types.

This function takes &mut self, so it can’t conveniently return a shared reference to the resulting TypeResolution: the shared reference would extend the mutable borrow, and you wouldn’t be able to use self for anything else. Instead, you should use register_type or one of resolve!, resolve_inner! or resolve_inner_binary!.

fn image_data( &mut self, image: Handle<Expression>, span: Span ) -> Result<(ImageClass, bool), Error<'source>>

fn prepare_args<'b>( &mut self, args: &'b [Handle<Expression<'source>>], min_args: u32, span: Span ) -> ArgumentContext<'b, 'source>

fn binary_op_splat( &mut self, op: BinaryOperator, left: &mut Handle<Expression>, right: &mut Handle<Expression> ) -> Result<(), Error<'source>>

Insert splats, if needed by the non-‘*’ operations.

See the “Binary arithmetic expressions with mixed scalar and vector operands” table in the WebGPU Shading Language specification for relevant operators.

Multiply is not handled here as backends are expected to handle vec*scalar operations, so inserting splats into the IR increases size needlessly.

fn interrupt_emitter( &mut self, expression: Expression, span: Span ) -> Result<Handle<Expression>, Error<'source>>

Add a single expression to the expression table that is not covered by self.emitter.

This is useful for CallResult and AtomicResult expressions, which should not be covered by Emit statements.

fn apply_load_rule( &mut self, expr: TypedExpression ) -> Result<Handle<Expression>, Error<'source>>

Apply the WGSL Load Rule to expr.

If expr is has type ref<SC, T, A>, perform a load to produce a value of type T. Otherwise, return expr unchanged.

fn format_typeinner(&self, inner: &TypeInner) -> String

fn format_type(&self, handle: Handle<Type>) -> String

fn format_type_resolution(&self, resolution: &TypeResolution) -> String

fn ensure_type_exists(&mut self, inner: TypeInner) -> Handle<Type>