naga::front::wgsl::parse::directive

Module enable_extension

enable …; extensions in WGSL.

The focal point of this module is the EnableExtension API.

Structs

EnableExtensions

Tracks the status of every enable-extension known to Naga.

Enums

EnableExtension

An enable-extension not guaranteed to be present in all environments.

ImplementedEnableExtension

A variant of EnableExtension::Implemented.

UnimplementedEnableExtension

A variant of EnableExtension::Unimplemented.