naga::ir

Struct MeshStageInfo

pub struct MeshStageInfo {
    pub topology: MeshOutputTopology,
    pub max_vertices: u32,
    pub max_vertices_override: Option<Handle<Expression>>,
    pub max_primitives: u32,
    pub max_primitives_override: Option<Handle<Expression>>,
    pub vertex_output_type: Handle<Type>,
    pub primitive_output_type: Handle<Type>,
    pub output_variable: Handle<GlobalVariable>,
}

Information specific to mesh shader entry points.

Fields
topology: MeshOutputTopology

The type of primitive outputted.

max_vertices: u32

The maximum number of vertices a mesh shader may output.

max_vertices_override: Option<Handle<Expression>>

If pipeline constants are used, the expressions that override max_vertices

max_primitives: u32

The maximum number of primitives a mesh shader may output.

max_primitives_override: Option<Handle<Expression>>

If pipeline constants are used, the expressions that override max_primitives

vertex_output_type: Handle<Type>

The type used by vertex outputs, i.e. what is passed to setVertex.

primitive_output_type: Handle<Type>

The type used by primitive outputs, i.e. what is passed to setPrimitive.

output_variable: Handle<GlobalVariable>

The global variable holding the outputted vertices, primitives, and counts