naga::proc::constant_evaluator
Function component_wise_float

fn component_wise_float<const N: usize, const M: usize, F>(
    eval: &mut ConstantEvaluator<'_>,
    span: Span,
    exprs: [Handle<Expression>; N],
    handler: F,
) -> Result<Handle<Expression>, ConstantEvaluatorError>
where
    Float<M>: Into<Expression>,
    F: FnMut(Float<N>) -> Result<Float<M>, ConstantEvaluatorError> + Clone,

Attempts to evaluate multiple exprs as a combined Float to pass to handler. If exprs are vectors of the same length, handler is called for each corresponding component of each vector.

handler’s output is registered as a new expression. If exprs are vectors of the same length, a new vector expression is registered, composed of each component emitted by handler.