naga::proc

Function min_max_float_representable_by

pub fn min_max_float_representable_by(
    float: Scalar,
    int: Scalar,
) -> (Literal, Literal)

Returns a tuple of crate::Literals representing the minimum and maximum float values exactly representable by the provided float and integer types. Panics if float is not one of F16, F32, or F64, or int is not one of I32, U32, I64, or U64.