naga::proc::index

Function oob_local_types

pub fn oob_local_types(
    module: &Module,
    function: &Function,
    info: &FunctionInfo,
    policies: BoundsCheckPolicies,
) -> FastHashSet<Handle<Type>>

Returns all the types which we need out-of-bounds locals for; that is, all of the types which the code might attempt to get an out-of-bounds pointer to, in which case we yield a pointer to the out-of-bounds local of the correct type.