naga::proc::namer

Enum NameKey

pub enum NameKey {

    Constant(Handle<Constant>),
    GlobalVariable(Handle<GlobalVariable>),
    Type(Handle<Type>),
    StructMember(Handle<Type>, u32),
    Function(Handle<Function>),
    FunctionArgument(Handle<Function>, u32),
    FunctionLocal(Handle<Function>, Handle<LocalVariable>),
    FunctionOobLocal(Handle<Function>, Handle<Type>),
    EntryPoint(EntryPointIndex),
    EntryPointLocal(EntryPointIndex, Handle<LocalVariable>),
    EntryPointArgument(EntryPointIndex, u32),
    EntryPointOobLocal(EntryPointIndex, Handle<Type>),
    ExternalTextureGlobalVariable(Handle<GlobalVariable>, ExternalTextureNameKey),
    ExternalTextureFunctionArgument(Handle<Function>, u32, ExternalTextureNameKey),

}

Variants

Constant(Handle<Constant>)

GlobalVariable(Handle<GlobalVariable>)

Type(Handle<Type>)

StructMember(Handle<Type>, u32)

Function(Handle<Function>)

FunctionArgument(Handle<Function>, u32)

FunctionLocal(Handle<Function>, Handle<LocalVariable>)

FunctionOobLocal(Handle<Function>, Handle<Type>)

A local variable used by ReadZeroSkipWrite bounds-check policy when it needs to produce a pointer-typed result for an OOB access. These are unique per accessed type, so the second element is a type handle. See docs for crate::back::msl.

EntryPoint(EntryPointIndex)

EntryPointLocal(EntryPointIndex, Handle<LocalVariable>)

EntryPointArgument(EntryPointIndex, u32)

EntryPointOobLocal(EntryPointIndex, Handle<Type>)

Entry point version of FunctionOobLocal.

ExternalTextureGlobalVariable(Handle<GlobalVariable>, ExternalTextureNameKey)

A global variable holding a component of a lowered external texture.

See ExternalTextureNameKey for details.

ExternalTextureFunctionArgument(Handle<Function>, u32, ExternalTextureNameKey)

A function argument holding a component of a lowered external texture.

See ExternalTextureNameKey for details.