naga::proc
Module overloads

Overload resolution for builtin functions.

This module defines the OverloadSet trait, which provides methods the validator and typifier can use to check the types to builtin functions, determine their result types, and produce diagnostics that explain why a given application is not allowed and suggest fixes.

You can call MathFunction::overloads to obtain an impl OverloadSet representing the given MathFunction’s overloads.

Re-exports

pub use rule::Conclusion;

pub use rule::MissingSpecialType;

pub use rule::Rule;

Modules

any_overload_set 🔒

Dynamically dispatched OverloadSets.

constructor_set 🔒

A set of type constructors, represented as a bitset.

list 🔒

An OverloadSet represented as a vector of rules.

mathfunction 🔒

Overload sets for ir::MathFunction.

one_bits_iter 🔒

An iterator over bitmasks.

regular 🔒

A representation for highly regular overload sets common in Naga IR.

rule 🔒

Type rules.

scalar_set 🔒

A set of scalar types, represented as a bitset.

utils 🔒

Utility functions for constructing List overload sets.

Traits

OverloadSet

A trait for types representing of a set of Naga IR type rules.