naga::proc::overloads::regular

Function make_rule

fn make_rule(
    arity: usize,
    size: ConstructorSize,
    scalar: Scalar,
    conclusion_rule: ConclusionRule,
) -> Rule

Construct a Regular member Rule for the given arity and type.

Regular can only represent rules where all the argument types and the return type are the same, so just knowing arity and inner is sufficient.