naga::proc

Module type_methods

Methods on or related to TypeInner, Scalar, ScalarKind, and VectorSize.

Macros

define_int_float_limits 🔒

Constants

POINTER_SPAN 🔒

Traits

IntFloatLimits

Helper trait for providing the min and max values exactly representable by the integer type Self and floating point type F.

Functions

concrete_int_scalars

Produce all concrete integer ir::Scalars.

min_max_float_representable_by

Returns a tuple of crate::Literals representing the minimum and maximum float values exactly representable by the provided float and integer types. Panics if float is not one of F16, F32, or F64, or int is not one of I32, U32, I64, or U64.

vector_size_str

Helper function that returns the string corresponding to the VectorSize

vector_sizes

Produce all vector sizes.