naga::span

Struct Span

pub struct Span {
    start: u32,
    end: u32,
}

A source code span, used for error reporting.

Fields
start: u32end: u32
Implementations

impl Span

pub const UNDEFINED: Self

pub const fn new(start: u32, end: u32) -> Self

Creates a new Span from a range of byte indices

Note: end is exclusive, it doesnโt belong to the Span

pub const fn until(&self, other: &Self) -> Self

Returns a new Span starting at self and ending at other

pub fn subsume(&mut self, other: Self)

Modifies self to contain the smallest Span possible that contains both self and other

pub fn total_span<T: Iterator<Item = Self>>(from: T) -> Self

Returns the smallest Span possible that contains all the Spans defined in the from iterator

pub fn to_range(self) -> Option<Range<usize>>

Converts self to a range if the span is not unknown

pub fn is_defined(&self) -> bool

Check whether self was defined or is a default/unknown span

pub fn location(&self, source: &str) -> SourceLocation

Return a SourceLocation for this span in the provided source.