Struct naga::EarlyDepthTest

pub struct EarlyDepthTest {
    pub conservative: Option<ConservativeDepth>,
}

Early fragment tests.

In a standard situation, if a driver determines that it is possible to switch on early depth test, it will.

Typical situations when early depth test is switched off:

Calling discard in a shader.

Writing to the depth buffer, unless ConservativeDepth is enabled.

To use in a shader:

GLSL: layout(early_fragment_tests) in;

HLSL: Attribute earlydepthstencil

SPIR-V: ExecutionMode EarlyFragmentTests

WGSL: @early_depth_test

For more, see:

https://www.khronos.org/opengl/wiki/Early_Fragment_Test#Explicit_specification

https://docs.microsoft.com/en-us/windows/win32/direct3dhlsl/sm5-attributes-earlydepthstencil

https://www.khronos.org/registry/SPIR-V/specs/unified1/SPIRV.html#Execution_Mode

Fields
conservative: Option<ConservativeDepth>