naga

Struct Handle

pub struct Handle<T> {
    index: NonMaxU32,
    marker: PhantomData<T>,
}

A strongly typed reference to an arena item.

A Handle value can be used as an index into an Arena or UniqueArena.

Fields
index: NonMaxU32marker: PhantomData<T>
Implementations

impl<T> Handle<T>

pub(crate) const fn new(index: NonMaxU32) -> Self

pub const fn index(self) -> usize

Returns the index of this handle.

pub(super) fn from_usize(index: usize) -> Self

Convert a usize index into a Handle<T>.

pub(super) const unsafe fn from_usize_unchecked(index: usize) -> Self

Convert a usize index into a Handle<T>, without range checks.

pub fn write_prefixed( &self, formatter: &mut Formatter<'_>, prefix: &'static str, ) -> Result

Write this handle’s index to formatter, preceded by prefix.

impl<T> Handle<T>

fn check_valid_for(self, arena: &Arena<T>) -> Result<(), InvalidHandleError>

Check that self is valid within arena using Arena::check_contains_handle.

fn check_valid_for_uniq( self, arena: &UniqueArena<T>, ) -> Result<(), InvalidHandleError>
where T: Eq + Hash,

Check that self is valid within arena using UniqueArena::check_contains_handle.

fn check_dep(self, depends_on: Self) -> Result<Self, FwdDepError>

Check that depends_on was constructed before self by comparing handle indices.

If self is a valid handle (i.e., it has been validated using Self::check_valid_for) and this function returns Ok, then it may be assumed that depends_on is also valid. In naga’s current arena-based implementation, this is useful for validating recursive definitions of arena-based values in linear time.

Errors

If depends_on’s handle is from the same Arena as self's, but not constructed earlier than self’s, this function returns an error.

fn check_dep_opt(self, depends_on: Option<Self>) -> Result<Self, FwdDepError>

Like Self::check_dep, except for Optional handle values.

fn check_dep_iter( self, depends_on: impl Iterator<Item = Self>, ) -> Result<Self, FwdDepError>

Like Self::check_dep, except for Iterators over handle values.