naga::valid

Struct Validator

pub struct Validator { /* private fields */ }

Implementations

impl Validator

pub fn validate_literal(&self, literal: Literal) -> Result<(), LiteralError>

impl Validator

pub fn new(flags: ValidationFlags, capabilities: Capabilities) -> Self

Create a validator for Naga Modules.

The flags argument indicates which stages of validation the returned Validator should perform. Skipping stages can make validation somewhat faster, but the validator may not reject some invalid modules. Regardless of flags, validation always returns a usable ModuleInfo value on success.

If flags contains everything in ValidationFlags::default(), then the returned Naga Validator will reject any Module that would use capabilities not included in capabilities.

pub const fn subgroup_stages(&mut self, stages: ShaderStages) -> &mut Self

pub const fn subgroup_operations( &mut self, operations: SubgroupOperationSet, ) -> &mut Self

pub fn reset(&mut self)

Reset the validator internals

pub fn validate( &mut self, module: &Module, ) -> Result<ModuleInfo, WithSpan<ValidationError>>

Check the given module to be valid.

pub fn validate_resolved_overrides( &mut self, module: &Module, ) -> Result<ModuleInfo, WithSpan<ValidationError>>

Check the given module to be valid, requiring overrides to be resolved.

This is the same as validate, except that any override whose value is not a fully-evaluated constant expression is treated as an error.