Module naga::back

Backend functions that export shader Modules into binary and text formats.

Modules

dot

Backend for DOT (Graphviz).

glsl

Backend for GLSL (OpenGL Shading Language).

hlsl

Backend for HLSL (High-Level Shading Language).

msl

Backend for MSL (Metal Shading Language).

pipeline_constants

spv

Backend for SPIR-V (Standard Portable Intermediate Representation).

wgsl

Backend for WGSL (WebGPU Shading Language).

Structs

FunctionCtx

Helper structure that stores data needed when writing the function

Level

Indentation level.

RayFlag

Ray flags, for a RayDesc’s flags field.

Enums

FunctionType

Whether we’re generating an entry point or a regular function.

RayIntersectionType

The intersection test to use for ray queries.

Constants

COMPONENTS

Names of vector components.

INDENT

Indent for backends.

Functions

binary_operation_str

Helper function that returns the string corresponding to the BinaryOperator

Type Aliases

NeedBakeExpressions

Expressions that need baking.

PipelineConstants

Specifies the values of pipeline-overridable constants in the shader module.