Module naga::back

Backend functions that export shader Modules into binary and text formats.

Modules

spv

Backend for SPIR-V (Standard Portable Intermediate Representation).

Structs

FunctionCtx 🔒

Helper structure that stores data needed when writing the function

Level 🔒

RayFlag

Ray flags, for a RayDesc’s flags field.

Enums

FunctionType 🔒

Whether we’re generating an entry point or a regular function.

RayIntersectionType 🔒

Constants

BAKE_PREFIX 🔒

COMPONENTS 🔒

INDENT 🔒

Functions

binary_operation_str 🔒

Helper function that returns the string corresponding to the BinaryOperator

vector_size_str 🔒

Helper function that returns the string corresponding to the VectorSize

Type Aliases

NeedBakeExpressions 🔒