naga::back

Module glsl

Backend for GLSL (OpenGL Shading Language).

The main structure is Writer, it maintains internal state that is used to output a Module into glsl

Supported versions

Core

330

400

410

420

430

450

ES

300

310

Modules

conv 🔒

Contains simple 1:1 conversion functions.

features 🔒

Contains the features related code and the features querying method

keywords 🔒

Contains a constant with a slice of all the reserved keywords RESERVED_KEYWORDS

writer 🔒

Contains the Writer type.

Structs

BindingMapSerialization 🔒

Features

Structure used to encode additions to GLSL that aren’t supported by all versions.

IdGenerator 🔒

Helper structure that generates a number

ImmediateItem

All information to bind a single uniform value to the shader.

Options

Configuration used in the Writer.

PipelineOptions

A subset of options meant to be changed per pipeline.

ReflectionInfo

Reflection info for texture mappings and uniforms.

TextureMapping

Mapping between a texture and its sampler, if it exists.

VaryingLocation

VaryingName 🔒

Helper wrapper used to get a name for a varying

VaryingOptions 🔒

Assorted options needed for generating varyings.

Writer

Writer responsible for all code generation.

WriterFlags

Configuration flags for the Writer.

Enums

BinaryOperation 🔒

Binary operation with a different logic on the GLSL side.

Error

A GLSL compilation error.

Version

A GLSL version.

Constants

CLAMPED_LOD_SUFFIX 🔒

The suffix of the variable that will hold the calculated clamped level of detail for bounds checking in ImageLoad

FIRST_INSTANCE_BINDING

FREXP_FUNCTION 🔒

MODF_FUNCTION 🔒

SUPPORTED_CORE_VERSIONS

List of supported core GLSL versions.

SUPPORTED_ES_VERSIONS

List of supported es GLSL versions.

Functions

deserialize_binding_map 🔒

is_value_init_supported 🔒

Type Aliases

BackendResult 🔒

Shorthand result used internally by the backend

BindingMap

Mapping between resources and bindings.