Module naga::back::glsl

OpenGL shading language backend

The main structure is Writer, it maintains internal state that is used to output a Module into glsl

Supported versions

Core

330

400

410

420

430

450

460

ES

300

310

Structs

Features

Structure used to encode a set of additions to glsl that aren’t supported by all versions

Options

Structure that contains the configuration used in the Writer

PipelineOptions

ReflectionInfo

Structure that contains a reflection info

TextureMapping

Structure that connects a texture to a sampler or not

Writer

Main structure of the glsl backend responsible for all code generation

WriterFlags

Enums

Error

A glsl compilation error.

Version

glsl version

Constants

SUPPORTED_CORE_VERSIONS

List of supported core glsl versions

SUPPORTED_ES_VERSIONS

List of supported es glsl versions

Type Definitions

BindingMap