naga::back::glsl

Module conv

Contains simple 1:1 conversion functions.

Structs

ScalarString 🔒

Structure returned by glsl_scalar

Functions

glsl_built_in 🔒

Helper function that returns the glsl variable name for a builtin

glsl_dimension 🔒

Helper function that returns the glsl dimension string of ImageDimension

glsl_interpolation 🔒

Helper function that returns the string corresponding to the glsl interpolation qualifier

glsl_sampling 🔒

Return the GLSL auxiliary qualifier for the given sampling value.

glsl_scalar 🔒

Helper function that returns scalar related strings

glsl_storage_format 🔒

Helper function that returns the glsl storage format string of StorageFormat

glsl_storage_qualifier 🔒

Helper function that returns the string corresponding to the address space