naga::back::hlsl
Module writer

Structs

BindingArraySamplerInfo 🔒

Information for how to generate a binding_array<sampler> access.

EntryPointBinding 🔒

Structure contains information required for generating wrapped structure of all entry points arguments

EntryPointInterface 🔒

EpStructMember 🔒

MatrixType 🔒

Enums

InterfaceKey 🔒

Io 🔒

Constants

ABS_FUNCTION 🔒

COMPARISON_SAMPLER_HEAP_VAR 🔒

DIV_FUNCTION 🔒

EXTRACT_BITS_FUNCTION 🔒

FREXP_FUNCTION 🔒

INSERT_BITS_FUNCTION 🔒

LOCATION_SEMANTIC 🔒

MODF_FUNCTION 🔒

MOD_FUNCTION 🔒

NEG_FUNCTION 🔒

SAMPLER_HEAP_VAR 🔒

SPECIAL_CBUF_TYPE 🔒

SPECIAL_CBUF_VAR 🔒

SPECIAL_FIRST_INSTANCE 🔒

SPECIAL_FIRST_VERTEX 🔒

SPECIAL_OTHER 🔒

Functions

get_inner_matrix_data 🔒

get_inner_matrix_of_global_uniform 🔒

Returns the matrix data if the access chain starting at base:

get_inner_matrix_of_struct_array_member 🔒

Returns the matrix data if the access chain starting at base:

is_subgroup_builtin_binding 🔒