naga::back::msl
Struct Options

pub struct Options {
    pub lang_version: (u8, u8),
    pub per_entry_point_map: EntryPointResourceMap,
    pub inline_samplers: Vec<InlineSampler>,
    pub spirv_cross_compatibility: bool,
    pub fake_missing_bindings: bool,
    pub bounds_check_policies: BoundsCheckPolicies,
    pub zero_initialize_workgroup_memory: bool,
    pub force_loop_bounding: bool,
}

Fields
lang_version: (u8, u8)

(Major, Minor) target version of the Metal Shading Language.

per_entry_point_map: EntryPointResourceMap

Map of entry-point resources, indexed by entry point function name, to slots.

inline_samplers: Vec<InlineSampler>

Samplers to be inlined into the code.

spirv_cross_compatibility: bool

Make it possible to link different stages via SPIRV-Cross.

fake_missing_bindings: bool

Donโt panic on missing bindings, instead generate invalid MSL.

bounds_check_policies: BoundsCheckPolicies

Bounds checking policies.

zero_initialize_workgroup_memory: bool

Should workgroup variables be zero initialized (by polyfilling)?

force_loop_bounding: bool

If set, loops will have code injected into them, forcing the compiler to think the number of iterations is bounded.

Implementations

impl Options

fn resolve_local_binding( &self, binding: &Binding, mode: LocationMode, ) -> Result<ResolvedBinding, Error>

fn get_entry_point_resources( &self, ep: &EntryPoint, ) -> Option<&EntryPointResources>

fn get_resource_binding_target( &self, ep: &EntryPoint, res_binding: &ResourceBinding, ) -> Option<&BindTarget>

fn resolve_resource_binding( &self, ep: &EntryPoint, res_binding: &ResourceBinding, ) -> Result<ResolvedBinding, EntryPointError>

fn resolve_push_constants( &self, ep: &EntryPoint, ) -> Result<ResolvedBinding, EntryPointError>

fn resolve_sizes_buffer( &self, ep: &EntryPoint, ) -> Result<ResolvedBinding, EntryPointError>