naga::back::msl::writer
Struct ExpressionContext

struct ExpressionContext<'a> {
    function: &'a Function,
    origin: FunctionOrigin,
    info: &'a FunctionInfo,
    module: &'a Module,
    mod_info: &'a ModuleInfo,
    pipeline_options: &'a PipelineOptions,
    lang_version: (u8, u8),
    policies: BoundsCheckPolicies,
    guarded_indices: HandleSet<Expression>,
    force_loop_bounding: bool,
}

Fields
function: &'a Functionorigin: FunctionOrigininfo: &'a FunctionInfomodule: &'a Modulemod_info: &'a ModuleInfopipeline_options: &'a PipelineOptionslang_version: (u8, u8)policies: BoundsCheckPoliciesguarded_indices: HandleSet<Expression>

The set of expressions used as indices in ReadZeroSkipWrite-policy accesses. These may need to be cached in temporary variables. See index::find_checked_indexes for details.

force_loop_bounding: bool

See Writer::gen_force_bounded_loop_statements for details.

Implementations

impl<'a> ExpressionContext<'a>

fn resolve_type(&self, handle: Handle<Expression>) -> &'a TypeInner

fn image_needs_lod(&self, image: Handle<Expression>) -> bool

Return true if calls to image’s read and write methods should supply a level of detail.

Only mipmapped images need to specify a level of detail. Since 1D textures cannot have mipmaps, MSL requires that the level argument to texture1d queries and accesses must be a constexpr 0. It’s easiest just to omit the level entirely for 1D textures.

fn choose_bounds_check_policy( &self, pointer: Handle<Expression>, ) -> BoundsCheckPolicy

fn access_needs_check( &self, base: Handle<Expression>, index: GuardedIndex, ) -> Option<IndexableLength>

fn get_packed_vec_kind(&self, expr_handle: Handle<Expression>) -> Option<Scalar>