naga::back
Module pipeline_constants

Enums

PipelineConstantError

Functions

adjust_block 🔒

Replace every expression handle in block with its counterpart given by new_pos.

adjust_expr 🔒

Replace every expression handle in expr with its counterpart given by new_pos.

adjust_stmt 🔒

Replace every expression handle in stmt with its counterpart given by new_pos.

filter_emits_in_block 🔒

Adjust Emit statements in block to skip needs_pre_emit expressions we have introduced.

map_value_to_literal 🔒

process_function 🔒

Replace all override expressions in function with fully-evaluated constants.

process_override 🔒

Add a Constant to module for the override old_h.

process_overrides

Replace all overrides in module with constants.

process_pending 🔒

process_workgroup_size_override 🔒