naga::back::spv

Struct Writer

pub struct Writer {

    physical_layout: PhysicalLayout,
    logical_layout: LogicalLayout,
    id_gen: IdGenerator,
    capabilities_available: Option<FastHashSet<Capability>>,
    capabilities_used: FastIndexSet<Capability>,
    extensions_used: FastIndexSet<&'static str>,
    debugs: Vec<Instruction>,
    annotations: Vec<Instruction>,
    flags: WriterFlags,
    bounds_check_policies: BoundsCheckPolicies,
    zero_initialize_workgroup_memory: ZeroInitializeWorkgroupMemoryMode,
    force_loop_bounding: bool,
    use_storage_input_output_16: bool,
    void_type: Word,
    lookup_type: FastHashMap<LookupType, Word>,
    lookup_function: FastHashMap<Handle<Function>, Word>,
    lookup_function_type: FastHashMap<LookupFunctionType, Word>,
    wrapped_functions: FastHashMap<WrappedFunction, Word>,
    constant_ids: HandleVec<Expression, Word>,
    cached_constants: FastHashMap<CachedConstant, Word>,
    global_variables: HandleVec<GlobalVariable, GlobalVariable>,
    fake_missing_bindings: bool,
    binding_map: BindingMap,
    saved_cached: CachedExpressions,
    gl450_ext_inst_id: Word,
    temp_list: Vec<Word>,
    ray_get_committed_intersection_function: Option<Word>,
    ray_get_candidate_intersection_function: Option<Word>,
    io_f16_polyfills: F16IoPolyfill,

}

Fields
physical_layout: PhysicalLayoutlogical_layout: LogicalLayoutid_gen: IdGeneratorcapabilities_available: Option<FastHashSet<Capability>>

The set of capabilities modules are permitted to use.

This is initialized from Options::capabilities.

capabilities_used: FastIndexSet<Capability>

The set of capabilities used by this module.

If capabilities_available is Some, then this is always a subset of that.

extensions_used: FastIndexSet<&'static str>

The set of spirv extensions used.

debugs: Vec<Instruction>annotations: Vec<Instruction>flags: WriterFlagsbounds_check_policies: BoundsCheckPolicieszero_initialize_workgroup_memory: ZeroInitializeWorkgroupMemoryModeforce_loop_bounding: booluse_storage_input_output_16: boolvoid_type: Wordlookup_type: FastHashMap<LookupType, Word>lookup_function: FastHashMap<Handle<Function>, Word>lookup_function_type: FastHashMap<LookupFunctionType, Word>wrapped_functions: FastHashMap<WrappedFunction, Word>

Operations which have been wrapped in a helper function. The value is the ID of the function, which should be called instead of emitting code for the operation directly.

constant_ids: HandleVec<Expression, Word>

Indexed by const-expression handle indexes

cached_constants: FastHashMap<CachedConstant, Word>global_variables: HandleVec<GlobalVariable, GlobalVariable>fake_missing_bindings: boolbinding_map: BindingMapsaved_cached: CachedExpressionsgl450_ext_inst_id: Wordtemp_list: Vec<Word>ray_get_committed_intersection_function: Option<Word>ray_get_candidate_intersection_function: Option<Word>io_f16_polyfills: F16IoPolyfill

F16 I/O polyfill manager for handling f16 input/output variables when StorageInputOutput16 capability is not available.

Implementations

impl Writer

fn write_epilogue_position_y_flip( &mut self, position_id: Word, body: &mut Vec<Instruction>, ) -> Result<(), Error>

fn write_epilogue_frag_depth_clamp( &mut self, frag_depth_id: Word, body: &mut Vec<Instruction>, ) -> Result<(), Error>

fn write_entry_point_return( &mut self, value_id: Word, ir_result: &FunctionResult, result_members: &[ResultMember], body: &mut Vec<Instruction>, ) -> Result<(), Error>

impl Writer

pub(super) fn write_ray_query_get_intersection_function( &mut self, is_committed: bool, ir_module: &Module, ) -> Word

impl Writer

pub fn new(options: &Options<'_>) -> Result<Self, Error>

pub fn set_options(&mut self, options: &Options<'_>) -> Result<(), Error>

pub const fn lang_version(&self) -> (u8, u8)

Returns (major, minor) of the SPIR-V language version.

fn reset(&mut self)

Reset Writer to its initial state, retaining any allocations.

Why not just implement Recyclable for Writer? By design, Recyclable::recycle requires ownership of the value, not just &mut; see the trait documentation. But we need to use this method from functions like Writer::write, which only have &mut Writer. Workarounds include unsafe code (core::ptr::read, then write, ugh) or something like a Default impl that returns an oddly-initialized Writer, which is worse.

pub(super) fn require_any( &mut self, what: &'static str, capabilities: &[Capability], ) -> Result<(), Error>

Indicate that the code requires any one of the listed capabilities.

If nothing in capabilities appears in the available capabilities specified in the Options from which this Writer was created, return an error. The what string is used in the error message to explain what provoked the requirement. (If no available capabilities were given, assume everything is available.)

The first acceptable capability will be added to this Writer’s capabilities_used table, and an OpCapability emitted for it in the result. For this reason, more specific capabilities should be listed before more general.

pub(super) fn require_all( &mut self, capabilities: &[Capability], ) -> Result<(), Capability>

Indicate that the code requires all of the listed capabilities.

If all entries of capabilities appear in the available capabilities specified in the Options from which this Writer was created (including the case where Options::capabilities is None), add them all to this Writer’s capabilities_used table, and return Ok(()). If at least one of the listed capabilities is not available, do not add anything to the capabilities_used table, and return the first unavailable requested capability, wrapped in Err().

This method is does not return an Error in case of failure because it may be used in cases where the caller can recover (e.g., with a polyfill) if the requested capabilities are not available. In this case, it would be unnecessary work to find all the unavailable requested capabilities, and to allocate a Vec for them, just so we could return an Error::MissingCapabilities).

pub(super) fn use_extension(&mut self, extension: &'static str)

Indicate that the code uses the given extension.

pub(super) fn get_type_id(&mut self, lookup_ty: LookupType) -> Word

pub(super) fn get_handle_type_id(&mut self, handle: Handle<Type>) -> Word

pub(super) fn get_expression_lookup_type( &mut self, tr: &TypeResolution, ) -> LookupType

pub(super) fn get_expression_type_id(&mut self, tr: &TypeResolution) -> Word

pub(super) fn get_localtype_id(&mut self, local: LocalType) -> Word

pub(super) fn get_pointer_type_id( &mut self, base: Word, class: StorageClass, ) -> Word

pub(super) fn get_handle_pointer_type_id( &mut self, base: Handle<Type>, class: StorageClass, ) -> Word

pub(super) fn get_ray_query_pointer_id(&mut self) -> Word

pub(super) fn get_resolution_pointer_id( &mut self, resolution: &TypeResolution, class: StorageClass, ) -> Word

Return a SPIR-V type for a pointer to resolution.

The given resolution must be one that we can represent either as a LocalType::Pointer or LocalType::LocalPointer.

pub(super) fn get_numeric_type_id(&mut self, numeric: NumericType) -> Word

pub(super) fn get_u32_type_id(&mut self) -> Word

pub(super) fn get_f32_type_id(&mut self) -> Word

pub(super) fn get_vec2u_type_id(&mut self) -> Word

pub(super) fn get_vec2f_type_id(&mut self) -> Word

pub(super) fn get_vec3u_type_id(&mut self) -> Word

pub(super) fn get_f32_pointer_type_id(&mut self, class: StorageClass) -> Word

pub(super) fn get_vec2u_pointer_type_id(&mut self, class: StorageClass) -> Word

pub(super) fn get_vec3u_pointer_type_id(&mut self, class: StorageClass) -> Word

pub(super) fn get_bool_type_id(&mut self) -> Word

pub(super) fn get_vec2_bool_type_id(&mut self) -> Word

pub(super) fn get_vec3_bool_type_id(&mut self) -> Word

pub(super) fn decorate( &mut self, id: Word, decoration: Decoration, operands: &[Word], )

fn localtype_from_inner(&mut self, inner: &TypeInner) -> Option<LocalType>

Return inner as a LocalType, if that’s possible.

If inner can be represented as a LocalType, return Some(local_type).

Otherwise, return None. In this case, the type must always be looked up using a LookupType::Handle.

fn resolve_resource_binding( &self, res_binding: &ResourceBinding, ) -> Result<BindingInfo, Error>

Resolve the BindingInfo for a crate::ResourceBinding from the provided Writer::binding_map.

If the specified resource is not present in the binding map this will return an error, unless Writer::fake_missing_bindings is set.

fn write_wrapped_functions( &mut self, ir_function: &Function, info: &FunctionInfo, ir_module: &Module, ) -> Result<(), Error>

Emits code for any wrapper functions required by the expressions in ir_function. The IDs of any emitted functions will be stored in Self::wrapped_functions.

fn write_wrapped_binary_op( &mut self, op: BinaryOperator, return_type: NumericType, left_type: &TypeResolution, right_type: &TypeResolution, ) -> Result<(), Error>

Write a SPIR-V function that performs the operator op with Naga IR semantics.

Define a function that performs an integer division or modulo operation, except that using a divisor of zero or causing signed overflow with a divisor of -1 returns the numerator unchanged, rather than exhibiting undefined behavior.

Store the generated function’s id in the wrapped_functions table.

The operator op must be either Divide or Modulo.

Panics

The return_type, left_type or right_type arguments must all be integer scalars or vectors. If not, this function panics.

fn write_function( &mut self, ir_function: &Function, info: &FunctionInfo, ir_module: &Module, interface: Option<FunctionInterface<'_>>, debug_info: &Option<DebugInfoInner<'_>>, ) -> Result<Word, Error>

fn write_execution_mode( &mut self, function_id: Word, mode: ExecutionMode, ) -> Result<(), Error>

fn write_entry_point( &mut self, entry_point: &EntryPoint, info: &FunctionInfo, ir_module: &Module, debug_info: &Option<DebugInfoInner<'_>>, ) -> Result<Instruction, Error>

fn make_scalar(&mut self, id: Word, scalar: Scalar) -> Instruction

fn request_type_capabilities(&mut self, inner: &TypeInner) -> Result<(), Error>

fn write_numeric_type_declaration_local( &mut self, id: Word, numeric: NumericType, )

fn write_type_declaration_local(&mut self, id: Word, local_ty: LocalType)

fn write_type_declaration_arena( &mut self, module: &Module, handle: Handle<Type>, ) -> Result<Word, Error>

fn request_image_format_capabilities( &mut self, format: ImageFormat, ) -> Result<(), Error>

pub(super) fn get_index_constant(&mut self, index: Word) -> Word

pub(super) fn get_constant_scalar_with( &mut self, value: u8, scalar: Scalar, ) -> Result<Word, Error>

pub(super) fn get_constant_scalar(&mut self, value: Literal) -> Word

fn write_constant_scalar( &mut self, id: Word, value: &Literal, debug_name: Option<&String>, )

pub(super) fn get_constant_composite( &mut self, ty: LookupType, constituent_ids: &[Word], ) -> Word

fn write_constant_composite( &mut self, id: Word, ty: LookupType, constituent_ids: &[Word], debug_name: Option<&String>, )

pub(super) fn get_constant_null(&mut self, type_id: Word) -> Word

pub(super) fn write_constant_null(&mut self, type_id: Word) -> Word

fn write_constant_expr( &mut self, handle: Handle<Expression>, ir_module: &Module, mod_info: &ModuleInfo, ) -> Result<Word, Error>

pub(super) fn write_control_barrier( &mut self, flags: Barrier, block: &mut Block, )

pub(super) fn write_memory_barrier(&mut self, flags: Barrier, block: &mut Block)

fn generate_workgroup_vars_init_block( &mut self, entry_id: Word, ir_module: &Module, info: &FunctionInfo, local_invocation_id: Option<Word>, interface: &mut FunctionInterface<'_>, function: &mut Function, ) -> Option<Word>

fn write_varying( &mut self, ir_module: &Module, stage: ShaderStage, class: StorageClass, debug_name: Option<&str>, ty: Handle<Type>, binding: &Binding, ) -> Result<Word, Error>

Generate an OpVariable for one value in an EntryPoint’s IO interface.

The Bindings of the arguments and result of an EntryPoint’s Function describe a SPIR-V shader interface. In SPIR-V, the interface is represented by global variables in the Input and Output storage classes, with decorations indicating which builtin or location each variable corresponds to.

This function emits a single global OpVariable for a single value from the interface, and adds appropriate decorations to indicate which builtin or location it represents, how it should be interpolated, and so on. The class argument gives the variable’s SPIR-V storage class, which should be either Input or Output.

pub(super) fn load_io_with_f16_polyfill( &mut self, body: &mut Vec<Instruction>, varying_id: Word, target_type_id: Word, ) -> Word

Load an IO variable, converting from f32 to f16 if polyfill is active. Returns the id of the loaded value matching target_type_id.

pub(super) fn store_io_with_f16_polyfill( &mut self, body: &mut Vec<Instruction>, varying_id: Word, value_id: Word, )

Store an IO variable, converting from f16 to f32 if polyfill is active.

fn write_global_variable( &mut self, ir_module: &Module, global_variable: &GlobalVariable, ) -> Result<Word, Error>

fn decorate_struct_member( &mut self, struct_id: Word, index: usize, member: &StructMember, arena: &UniqueArena<Type>, ) -> Result<(), Error>

Write the necessary decorations for a struct member.

Emit decorations for the index’th member of the struct type designated by struct_id, described by member.

pub(super) fn get_function_type( &mut self, lookup_function_type: LookupFunctionType, ) -> Word

fn write_physical_layout(&mut self)

fn write_logical_layout( &mut self, ir_module: &Module, mod_info: &ModuleInfo, ep_index: Option<usize>, debug_info: &Option<DebugInfo<'_>>, ) -> Result<(), Error>

pub fn write( &mut self, ir_module: &Module, info: &ModuleInfo, pipeline_options: Option<&PipelineOptions>, debug_info: &Option<DebugInfo<'_>>, words: &mut Vec<Word>, ) -> Result<(), Error>

pub const fn get_capabilities_used(&self) -> &FastIndexSet<Capability>

Return the set of capabilities the last module written used.

pub fn decorate_non_uniform_binding_array_access( &mut self, id: Word, ) -> Result<(), Error>

pub(super) fn needs_f16_polyfill(&self, ty_inner: &TypeInner) -> bool