naga::common::wgsl
Trait TypeContext

pub trait TypeContext<W: Write> {
    // Required methods
    fn lookup_type(&self, handle: Handle<Type>) -> &Type;
    fn type_name(&self, handle: Handle<Type>) -> &str;
    fn write_override(&self, override: Handle<Override>, out: &mut W) -> Result;

    // Provided methods
    fn write_type(&self, handle: Handle<Type>, out: &mut W) -> Result { ... }
    fn write_type_inner(&self, inner: &TypeInner, out: &mut W) -> Result { ... }
}

A context for printing Naga IR types as WGSL.

This trait’s default methods write_type and write_type_inner do the work of formatting types as WGSL. Implementors must provide the remaining methods, to customize behavior for the context at hand.

For example, the WGSL backend would provide an implementation of type_name that handles hygienic renaming, whereas the WGSL front end would simply show the name that was given in the source.

Required Methods

fn lookup_type(&self, handle: Handle<Type>) -> &Type

Return the Type referred to by handle.

fn type_name(&self, handle: Handle<Type>) -> &str

Return the name to be used for the type referred to by handle.

fn write_override(&self, override: Handle<Override>, out: &mut W) -> Result

Write the WGSL form of override to out.

Provided Methods

fn write_type(&self, handle: Handle<Type>, out: &mut W) -> Result

Write the type ty as it would appear in a value’s declaration.

Write the type referred to by ty in module as it would appear in a var, let, etc. declaration, or in a function’s argument list.

fn write_type_inner(&self, inner: &TypeInner, out: &mut W) -> Result

Write the TypeInner inner as it would appear in a value’s declaration.

Write inner as it would appear in a var, let, etc. declaration, or in a function’s argument list.

Note that this cannot handle writing Struct types: those must be referred to by name, but the name isn’t available in TypeInner.