Enum naga::Expression

pub enum Expression {

    Access {
        base: Handle<Expression>,
        index: Handle<Expression>,
    },
    AccessIndex {
        base: Handle<Expression>,
        index: u32,
    },
    Constant(Handle<Constant>),
    Splat {
        size: VectorSize,
        value: Handle<Expression>,
    },
    Swizzle {
        size: VectorSize,
        vector: Handle<Expression>,
        pattern: [SwizzleComponent; 4],
    },
    Compose {
        ty: Handle<Type>,
        components: Vec<Handle<Expression>>,
    },
    FunctionArgument(u32),
    GlobalVariable(Handle<GlobalVariable>),
    LocalVariable(Handle<LocalVariable>),
    Load {
        pointer: Handle<Expression>,
    },
    ImageSample {
        image: Handle<Expression>,
        sampler: Handle<Expression>,
        coordinate: Handle<Expression>,
        array_index: Option<Handle<Expression>>,
        offset: Option<Handle<Constant>>,
        level: SampleLevel,
        depth_ref: Option<Handle<Expression>>,
    },
    ImageLoad {
        image: Handle<Expression>,
        coordinate: Handle<Expression>,
        array_index: Option<Handle<Expression>>,
        index: Option<Handle<Expression>>,
    },
    ImageQuery {
        image: Handle<Expression>,
        query: ImageQuery,
    },
    Unary {
        op: UnaryOperator,
        expr: Handle<Expression>,
    },
    Binary {
        op: BinaryOperator,
        left: Handle<Expression>,
        right: Handle<Expression>,
    },
    Select {
        condition: Handle<Expression>,
        accept: Handle<Expression>,
        reject: Handle<Expression>,
    },
    Derivative {
        axis: DerivativeAxis,
        expr: Handle<Expression>,
    },
    Relational {
        fun: RelationalFunction,
        argument: Handle<Expression>,
    },
    Math {
        fun: MathFunction,
        arg: Handle<Expression>,
        arg1: Option<Handle<Expression>>,
        arg2: Option<Handle<Expression>>,
    },
    As {
        expr: Handle<Expression>,
        kind: ScalarKind,
        convert: Option<Bytes>,
    },
    Call(Handle<Function>),
    ArrayLength(Handle<Expression>),

}

An expression that can be evaluated to obtain a value.

This is a Single Static Assignment (SSA) scheme similar to SPIR-V.

Variants

Access

Array access with a computed index.

Fields of Access

base: Handle<Expression>index: Handle<Expression>

AccessIndex

Array access with a known index.

Fields of AccessIndex

base: Handle<Expression>index: u32

Constant(Handle<Constant>)

Constant value.

Splat

Splat scalar into a vector.

Fields of Splat

size: VectorSizevalue: Handle<Expression>

Swizzle

Vector swizzle.

Fields of Swizzle

size: VectorSizevector: Handle<Expression>pattern: [SwizzleComponent; 4]

Compose

Composite expression.

Fields of Compose

ty: Handle<Type>components: Vec<Handle<Expression>>

FunctionArgument(u32)

Reference a function parameter, by its index.

GlobalVariable(Handle<GlobalVariable>)

Reference a global variable.

LocalVariable(Handle<LocalVariable>)

Reference a local variable.

Load

Load a value indirectly.

Fields of Load

pointer: Handle<Expression>

ImageSample

Sample a point from a sampled or a depth image.

Fields of ImageSample

image: Handle<Expression>sampler: Handle<Expression>coordinate: Handle<Expression>array_index: Option<Handle<Expression>>offset: Option<Handle<Constant>>level: SampleLeveldepth_ref: Option<Handle<Expression>>

ImageLoad

Load a texel from an image.

Fields of ImageLoad

image: Handle<Expression>coordinate: Handle<Expression>array_index: Option<Handle<Expression>>index: Option<Handle<Expression>>

For storage images, this is None. For sampled images, this is the Some(Level). For multisampled images, this is Some(Sample).

ImageQuery

Query information from an image.

Fields of ImageQuery

image: Handle<Expression>query: ImageQuery

Unary

Apply an unary operator.

Fields of Unary

op: UnaryOperatorexpr: Handle<Expression>

Binary

Apply a binary operator.

Fields of Binary

op: BinaryOperatorleft: Handle<Expression>right: Handle<Expression>

Select

Select between two values based on a condition.

Fields of Select

condition: Handle<Expression>

Boolean expression

accept: Handle<Expression>reject: Handle<Expression>

Derivative

Compute the derivative on an axis.

Fields of Derivative

axis: DerivativeAxisexpr: Handle<Expression>

Relational

Call a relational function.

Fields of Relational

fun: RelationalFunctionargument: Handle<Expression>

Math

Call a math function

Fields of Math

fun: MathFunctionarg: Handle<Expression>arg1: Option<Handle<Expression>>arg2: Option<Handle<Expression>>

As

Cast a simply type to another kind.

Fields of As

expr: Handle<Expression>

Source expression, which can only be a scalar or a vector.

kind: ScalarKind

Target scalar kind.

convert: Option<Bytes>

If provided, converts to the specified byte width. Otherwise, bitcast.

Call(Handle<Function>)

Result of calling another function.

ArrayLength(Handle<Expression>)

Get the length of an array. The expression must resolve to a pointer to an array with a dynamic size.

This doesn’t match the semantics of spirv’s OpArrayLength, which must be passed a pointer to a structure containing a runtime array in its’ last field.

Implementations

impl Expression

pub fn needs_pre_emit(&self) -> bool

Returns true if the expression is considered emitted at the start of a function.