Enum naga::Expression

pub enum Expression {

    Access {
        base: Handle<Expression>,
        index: Handle<Expression>,
    },
    AccessIndex {
        base: Handle<Expression>,
        index: u32,
    },
    Constant(Handle<Constant>),
    Compose {
        ty: Handle<Type>,
        components: Vec<Handle<Expression>>,
    },
    FunctionParameter(u32),
    GlobalVariable(Handle<GlobalVariable>),
    LocalVariable(Handle<LocalVariable>),
    Load {
        pointer: Handle<Expression>,
    },
    ImageSample {
        image: Handle<Expression>,
        sampler: Handle<Expression>,
        coordinate: Handle<Expression>,
        level: SampleLevel,
        depth_ref: Option<Handle<Expression>>,
    },
    ImageLoad {
        image: Handle<Expression>,
        coordinate: Handle<Expression>,
        index: Handle<Expression>,
    },
    Unary {
        op: UnaryOperator,
        expr: Handle<Expression>,
    },
    Binary {
        op: BinaryOperator,
        left: Handle<Expression>,
        right: Handle<Expression>,
    },
    Intrinsic {
        fun: IntrinsicFunction,
        argument: Handle<Expression>,
    },
    Transpose(Handle<Expression>),
    DotProduct(Handle<Expression>, Handle<Expression>),
    CrossProduct(Handle<Expression>, Handle<Expression>),
    As {
        expr: Handle<Expression>,
        kind: ScalarKind,
        convert: bool,
    },
    Derivative {
        axis: DerivativeAxis,
        expr: Handle<Expression>,
    },
    Call {
        origin: FunctionOrigin,
        arguments: Vec<Handle<Expression>>,
    },

}

An expression that can be evaluated to obtain a value.

Variants

Access

Fields

base: Handle<Expression>

index: Handle<Expression>

Array access with a computed index.

AccessIndex

Fields

base: Handle<Expression>

index: u32

Array access with a known index.

Constant(Handle<Constant>)

Constant value.

Compose

Fields

ty: Handle<Type>

components: Vec<Handle<Expression>>

Composite expression.

FunctionParameter(u32)

Reference a function parameter, by its index.

GlobalVariable(Handle<GlobalVariable>)

Reference a global variable.

LocalVariable(Handle<LocalVariable>)

Reference a local variable.

Load

Fields

pointer: Handle<Expression>

Load a value indirectly.

ImageSample

Fields

image: Handle<Expression>

sampler: Handle<Expression>

coordinate: Handle<Expression>

level: SampleLevel

depth_ref: Option<Handle<Expression>>

Sample a point from a sampled or a depth image.

ImageLoad

Fields

image: Handle<Expression>

coordinate: Handle<Expression>

index: Handle<Expression>

This is either LOD index, or sample index, based on the image class.

Load a texel from an image.

Unary

Fields

op: UnaryOperator

expr: Handle<Expression>

Apply an unary operator.

Binary

Fields

op: BinaryOperator

left: Handle<Expression>

right: Handle<Expression>

Apply a binary operator.

Intrinsic

Fields

fun: IntrinsicFunction

argument: Handle<Expression>

Call an intrinsic function.

Transpose(Handle<Expression>)

Transpose of a matrix.

DotProduct(Handle<Expression>, Handle<Expression>)

Dot product between two vectors.

CrossProduct(Handle<Expression>, Handle<Expression>)

Cross product between two vectors.

As

Fields

expr: Handle<Expression>

Source expression, which can only be a scalar or a vector.

kind: ScalarKind

Target scalar kind.

convert: bool

True = conversion needs to take place; False = bitcast.

Cast a simply type to another kind.

Derivative

Fields

axis: DerivativeAxis

expr: Handle<Expression>

Compute the derivative on an axis.

Call

Fields

origin: FunctionOrigin

arguments: Vec<Handle<Expression>>

Call another function.