Enum naga::Expression

pub enum Expression {

    Access {
        base: Handle<Expression>,
        index: Handle<Expression>,
    },
    AccessIndex {
        base: Handle<Expression>,
        index: u32,
    },
    Constant(Handle<Constant>),
    Compose {
        ty: Handle<Type>,
        components: Vec<Handle<Expression>>,
    },
    FunctionParameter(u32),
    GlobalVariable(Handle<GlobalVariable>),
    LocalVariable(Handle<LocalVariable>),
    Load {
        pointer: Handle<Expression>,
    },
    ImageSample {
        image: Handle<Expression>,
        sampler: Handle<Expression>,
        coordinate: Handle<Expression>,
        level: SampleLevel,
        depth_ref: Option<Handle<Expression>>,
    },
    ImageLoad {
        image: Handle<Expression>,
        coordinate: Handle<Expression>,
        index: Handle<Expression>,
    },
    Unary {
        op: UnaryOperator,
        expr: Handle<Expression>,
    },
    Binary {
        op: BinaryOperator,
        left: Handle<Expression>,
        right: Handle<Expression>,
    },
    Intrinsic {
        fun: IntrinsicFunction,
        argument: Handle<Expression>,
    },
    Transpose(Handle<Expression>),
    DotProduct(Handle<Expression>, Handle<Expression>),
    CrossProduct(Handle<Expression>, Handle<Expression>),
    As {
        expr: Handle<Expression>,
        kind: ScalarKind,
        convert: bool,
    },
    Derivative {
        axis: DerivativeAxis,
        expr: Handle<Expression>,
    },
    Call {
        origin: FunctionOrigin,
        arguments: Vec<Handle<Expression>>,
    },

}

An expression that can be evaluated to obtain a value.

Variants

Access

Array access with a computed index.

Fields of Access

base: Handle<Expression>index: Handle<Expression>

AccessIndex

Array access with a known index.

Fields of AccessIndex

base: Handle<Expression>index: u32

Constant

Constant value.

Tuple Fields of Constant

0: Handle<Constant>

Compose

Composite expression.

Fields of Compose

ty: Handle<Type>components: Vec<Handle<Expression>>

FunctionParameter

Reference a function parameter, by its index.

Tuple Fields of FunctionParameter

0: u32

GlobalVariable

Reference a global variable.

Tuple Fields of GlobalVariable

0: Handle<GlobalVariable>

LocalVariable

Reference a local variable.

Tuple Fields of LocalVariable

0: Handle<LocalVariable>

Load

Load a value indirectly.

Fields of Load

pointer: Handle<Expression>

ImageSample

Sample a point from a sampled or a depth image.

Fields of ImageSample

image: Handle<Expression>sampler: Handle<Expression>coordinate: Handle<Expression>level: SampleLeveldepth_ref: Option<Handle<Expression>>

ImageLoad

Load a texel from an image.

Fields of ImageLoad

image: Handle<Expression>coordinate: Handle<Expression>index: Handle<Expression>

This is either LOD index, or sample index, based on the image class.

Unary

Apply an unary operator.

Fields of Unary

op: UnaryOperatorexpr: Handle<Expression>

Binary

Apply a binary operator.

Fields of Binary

op: BinaryOperatorleft: Handle<Expression>right: Handle<Expression>

Intrinsic

Call an intrinsic function.

Fields of Intrinsic

fun: IntrinsicFunctionargument: Handle<Expression>

Transpose

Transpose of a matrix.

Tuple Fields of Transpose

0: Handle<Expression>

DotProduct

Dot product between two vectors.

Tuple Fields of DotProduct

0: Handle<Expression>1: Handle<Expression>

CrossProduct

Cross product between two vectors.

Tuple Fields of CrossProduct

0: Handle<Expression>1: Handle<Expression>

As

Cast a simply type to another kind.

Fields of As

expr: Handle<Expression>

Source expression, which can only be a scalar or a vector.

kind: ScalarKind

Target scalar kind.

convert: bool

True = conversion needs to take place; False = bitcast.

Derivative

Compute the derivative on an axis.

Fields of Derivative

axis: DerivativeAxisexpr: Handle<Expression>

Call

Call another function.

Fields of Call

origin: FunctionOriginarguments: Vec<Handle<Expression>>