Enum naga::Statement

pub enum Statement {
    Emit(Range<Expression>),
    Block(Block),
    If {
        condition: Handle<Expression>,
        accept: Block,
        reject: Block,
    },
    Switch {
        selector: Handle<Expression>,
        cases: Vec<SwitchCase>,
        default: Block,
    },
    Loop {
        body: Block,
        continuing: Block,
    },
    Break,
    Continue,
    Return {
        value: Option<Handle<Expression>>,
    },
    Kill,
    Store {
        pointer: Handle<Expression>,
        value: Handle<Expression>,
    },
    ImageStore {
        image: Handle<Expression>,
        coordinate: Handle<Expression>,
        array_index: Option<Handle<Expression>>,
        value: Handle<Expression>,
    },
    Call {
        function: Handle<Function>,
        arguments: Vec<Handle<Expression>>,
        result: Option<Handle<Expression>>,
    },
}

Instructions which make up an executable block.

Variants

Emit(Range<Expression>)

Emit a range of expressions, visible to all statements that follow in this block.

Block(Block)

A block containing more statements, to be executed sequentially.

If

Conditionally executes one of two blocks, based on the value of the condition.

Fields of If

condition: Handle<Expression>accept: Blockreject: Block

Switch

Conditionally executes one of multiple blocks, based on the value of the selector.

Fields of Switch

selector: Handle<Expression>cases: Vec<SwitchCase>default: Block

Loop

Executes a block repeatedly.

Fields of Loop

body: Blockcontinuing: Block

Break

Exits the loop.

Continue

Skips execution to the next iteration of the loop.

Return

Returns from the function (possibly with a value).

Fields of Return

value: Option<Handle<Expression>>

Kill

Aborts the current shader execution.

Store

Stores a value at an address.

This statement is a barrier for any operations on the Expression::LocalVariable or Expression::GlobalVariable that is the destination of an access chain, started from the pointer.

Fields of Store

pointer: Handle<Expression>value: Handle<Expression>

ImageStore

Stores a value to an image.

Image has to point into a global variable of type TypeInner::Image. This statement is a barrier for any operations on the corresponding Expression::GlobalVariable for this image.

Fields of ImageStore

image: Handle<Expression>coordinate: Handle<Expression>array_index: Option<Handle<Expression>>value: Handle<Expression>

Call

Calls a function.

If the result is Some, the corresponding expression has to be Expression::Call, and this statement serves as a barrier for any operations on that expression.

Fields of Call

function: Handle<Function>arguments: Vec<Handle<Expression>>result: Option<Handle<Expression>>