Enum naga::TypeInner

pub enum TypeInner {
    Scalar {
        kind: ScalarKind,
        width: Bytes,
    },
    Vector {
        size: VectorSize,
        kind: ScalarKind,
        width: Bytes,
    },
    Matrix {
        columns: VectorSize,
        rows: VectorSize,
        width: Bytes,
    },
    Pointer {
        base: Handle<Type>,
        class: StorageClass,
    },
    ValuePointer {
        size: Option<VectorSize>,
        kind: ScalarKind,
        width: Bytes,
        class: StorageClass,
    },
    Array {
        base: Handle<Type>,
        size: ArraySize,
        stride: u32,
    },
    Struct {
        top_level: bool,
        members: Vec<StructMember>,
        span: u32,
    },
    Image {
        dim: ImageDimension,
        arrayed: bool,
        class: ImageClass,
    },
    Sampler {
        comparison: bool,
    },
}

Enum with additional information, depending on the kind of type.

Variants

Scalar

Number of integral or floating-point kind.

Fields of Scalar

kind: ScalarKindwidth: Bytes

Vector

Vector of numbers.

Fields of Vector

size: VectorSizekind: ScalarKindwidth: Bytes

Matrix

Matrix of floats.

Fields of Matrix

columns: VectorSizerows: VectorSizewidth: Bytes

Pointer

Pointer to another type.

Fields of Pointer

base: Handle<Type>class: StorageClass

ValuePointer

Pointer to a value.

Fields of ValuePointer

size: Option<VectorSize>kind: ScalarKindwidth: Bytesclass: StorageClass

Array

Homogenous list of elements.

Fields of Array

base: Handle<Type>size: ArraySizestride: u32

Struct

User-defined structure.

Fields of Struct

top_level: boolmembers: Vec<StructMember>span: u32

Image

Possibly multidimensional array of texels.

Fields of Image

dim: ImageDimensionarrayed: boolclass: ImageClass

Sampler

Can be used to sample values from images.

Fields of Sampler

comparison: bool

Implementations

impl TypeInner

pub fn to_wgsl(
&self,
types: &Arena<Type>,
constants: &Arena<Constant>
) -> String

Formats the type as it is written in wgsl.

For example vec3<f32>.

Note: The names of a TypeInner::Struct is not known. Therefore this method will simply return “struct” for them.

impl TypeInner

pub fn scalar_kind(&self) -> Option<ScalarKind>

pub fn span(&self, constants: &Arena<Constant>) -> u32