naga
Module front

Frontend parsers that consume binary and text shaders and load them into Modules.

Modules

atomic_upgrade

Upgrade the types of scalars observed to be accessed as atomics to Atomic types.

glsl

Frontend for GLSL (OpenGL Shading Language).

interpolator 🔒

Interpolation defaults.

spv

Frontend for SPIR-V (Standard Portable Intermediate Representation).

type_gen 🔒

Type generators.

wgsl

Frontend for WGSL (WebGPU Shading Language).

Structs

SymbolTable

Structure responsible for managing variable lookups and keeping track of lexical scopes

Typifier

A table of types for an Arena<Expression>.

Type Aliases

Scope 🔒

Type representing a lexical scope, associating a name to a single variable