naga::front::glsl::builtins
Function texture_args_generator

fn texture_args_generator(
    options: TextureArgsOptions,
    f: impl FnMut(ScalarKind, ImageDimension, bool, bool, bool),
)

Helper function to generate the image components for texture/image builtins

Calls the passed function f with:

f(ScalarKind, ImageDimension, arrayed, multi, shadow)

options controls extra image variants generation like multisampling and depth, see the struct documentation