naga::front::glsl
Enum TokenValue

pub enum TokenValue {

    Identifier(String),
    FloatConstant(Float),
    IntConstant(Integer),
    BoolConstant(bool),
    Layout,
    In,
    Out,
    InOut,
    Uniform,
    Buffer,
    Const,
    Shared,
    Restrict,
    MemoryQualifier(StorageAccess),
    Invariant,
    Interpolation(Interpolation),
    Sampling(Sampling),
    Precision,
    PrecisionQualifier(Precision),
    Continue,
    Break,
    Return,
    Discard,
    If,
    Else,
    Switch,
    Case,
    Default,
    While,
    Do,
    For,
    Void,
    Struct,
    TypeName(Type),
    Assign,
    AddAssign,
    SubAssign,
    MulAssign,
    DivAssign,
    ModAssign,
    LeftShiftAssign,
    RightShiftAssign,
    AndAssign,
    XorAssign,
    OrAssign,
    Increment,
    Decrement,
    LogicalOr,
    LogicalAnd,
    LogicalXor,
    LessEqual,
    GreaterEqual,
    Equal,
    NotEqual,
    LeftShift,
    RightShift,
    LeftBrace,
    RightBrace,
    LeftParen,
    RightParen,
    LeftBracket,
    RightBracket,
    LeftAngle,
    RightAngle,
    Comma,
    Semicolon,
    Colon,
    Dot,
    Bang,
    Dash,
    Tilde,
    Plus,
    Star,
    Slash,
    Percent,
    VerticalBar,
    Caret,
    Ampersand,
    Question,

}

A token passed from the lexing used in the parsing.

This type is exported since itโs returned in the InvalidToken error.

Variants

Identifier(String)

FloatConstant(Float)

IntConstant(Integer)

BoolConstant(bool)

Layout

In

Out

InOut

Uniform

Buffer

Const

Shared

Restrict

MemoryQualifier(StorageAccess)

A glsl memory qualifier such as writeonly

The associated crate::StorageAccess is the access being allowed (for example writeonly has an associated value of crate::StorageAccess::STORE)

Invariant

Interpolation(Interpolation)

Sampling(Sampling)

Precision

PrecisionQualifier(Precision)

Continue

Break

Return

Discard

If

Else

Switch

Case

Default

While

Do

For

Void

Struct

TypeName(Type)

Assign

AddAssign

SubAssign

MulAssign

DivAssign

ModAssign

LeftShiftAssign

RightShiftAssign

AndAssign

XorAssign

OrAssign

Increment

Decrement

LogicalOr

LogicalAnd

LogicalXor

LessEqual

GreaterEqual

Equal

NotEqual

LeftShift

RightShift

LeftBrace

RightBrace

LeftParen

RightParen

LeftBracket

RightBracket

LeftAngle

RightAngle

Comma

Semicolon

Colon

Dot

Bang

Dash

Tilde

Plus

Star

Slash

Percent

VerticalBar

Caret

Ampersand

Question