Module naga::front::spv

SPIR-V frontend

ID lookups

Our IR links to everything with Handle, while SPIR-V uses IDs. In order to keep track of the associations, the parser has many lookup tables. There map spv::Word into a specific IR handle, plus potentially a bit of extra info, such as the related SPIR-V type ID. TODO: would be nice to find ways that avoid looking up as much

Structs

Parser

SamplingFlags
Flags describing sampling method.

Enums

Error

ModuleState

Constants

SUPPORTED_CAPABILITIES

SUPPORTED_EXTENSIONS

SUPPORTED_EXT_SETS

Functions

parse_u8_slice