Enum naga::front::wgsl::error::Error

pub(crate) enum Error<'a> {

    Unexpected(Span, ExpectedToken<'a>),
    UnexpectedComponents(Span),
    UnexpectedOperationInConstContext(Span),
    BadNumber(Span, NumberError),
    BadMatrixScalarKind(Span, Scalar),
    BadAccessor(Span),
    BadTexture(Span),
    BadTypeCast {
        span: Span,
        from_type: Box<str>,
        to_type: Box<str>,
    },
    BadTextureSampleType {
        span: Span,
        scalar: Scalar,
    },
    BadIncrDecrReferenceType(Span),
    InvalidResolve(ResolveError),
    InvalidForInitializer(Span),
    InvalidBreakIf(Span),
    InvalidGatherComponent(Span),
    InvalidConstructorComponentType(Span, i32),
    InvalidIdentifierUnderscore(Span),
    ReservedIdentifierPrefix(Span),
    UnknownAddressSpace(Span),
    RepeatedAttribute(Span),
    UnknownAttribute(Span),
    UnknownBuiltin(Span),
    UnknownAccess(Span),
    UnknownIdent(Span, &'a str),
    UnknownScalarType(Span),
    UnknownType(Span),
    UnknownStorageFormat(Span),
    UnknownConservativeDepth(Span),
    SizeAttributeTooLow(Span, u32),
    AlignAttributeTooLow(Span, Alignment),
    NonPowerOfTwoAlignAttribute(Span),
    InconsistentBinding(Span),
    TypeNotConstructible(Span),
    TypeNotInferable(Span),
    InitializationTypeMismatch {
        name: Span,
        expected: Box<str>,
        got: Box<str>,
    },
    DeclMissingTypeAndInit(Span),
    MissingAttribute(&'static str, Span),
    InvalidAtomicPointer(Span),
    InvalidAtomicOperandType(Span),
    InvalidRayQueryPointer(Span),
    Pointer(&'static str, Span),
    NotPointer(Span),
    NotReference(&'static str, Span),
    InvalidAssignment {
        span: Span,
        ty: InvalidAssignmentType,
    },
    ReservedKeyword(Span),
    Redefinition {
        previous: Span,
        current: Span,
    },
    RecursiveDeclaration {
        ident: Span,
        usage: Span,
    },
    CyclicDeclaration {
        ident: Span,
        path: Box<[(Span, Span)]>,
    },
    InvalidSwitchValue {
        uint: bool,
        span: Span,
    },
    CalledEntryPoint(Span),
    WrongArgumentCount {
        span: Span,
        expected: Range<u32>,
        found: u32,
    },
    FunctionReturnsVoid(Span),
    InvalidWorkGroupUniformLoad(Span),
    Internal(&'static str),
    ExpectedConstExprConcreteIntegerScalar(Span),
    ExpectedNonNegative(Span),
    ExpectedPositiveArrayLength(Span),
    MissingWorkgroupSize(Span),
    ConstantEvaluatorError(Box<ConstantEvaluatorError>, Span),
    AutoConversion(Box<AutoConversionError>),
    AutoConversionLeafScalar(Box<AutoConversionLeafScalarError>),
    ConcretizationFailed(Box<ConcretizationFailedError>),
    ExceededLimitForNestedBraces {
        span: Span,
        limit: u8,
    },
    PipelineConstantIDValue(Span),
    NotBool(Span),
    ConstAssertFailed(Span),

}

Variants

Unexpected(Span, ExpectedToken<'a>)

UnexpectedComponents(Span)

UnexpectedOperationInConstContext(Span)

BadNumber(Span, NumberError)

BadMatrixScalarKind(Span, Scalar)

BadAccessor(Span)

BadTexture(Span)

BadTypeCast

Fields

span: Span

from_type: Box<str>

to_type: Box<str>

BadTextureSampleType

Fields

span: Span

scalar: Scalar

BadIncrDecrReferenceType(Span)

InvalidResolve(ResolveError)

InvalidForInitializer(Span)

InvalidBreakIf(Span)

A break if appeared outside of a continuing block

InvalidGatherComponent(Span)

InvalidConstructorComponentType(Span, i32)

InvalidIdentifierUnderscore(Span)

ReservedIdentifierPrefix(Span)

UnknownAddressSpace(Span)

RepeatedAttribute(Span)

UnknownAttribute(Span)

UnknownBuiltin(Span)

UnknownAccess(Span)

UnknownIdent(Span, &'a str)

UnknownScalarType(Span)

UnknownType(Span)

UnknownStorageFormat(Span)

UnknownConservativeDepth(Span)

SizeAttributeTooLow(Span, u32)

AlignAttributeTooLow(Span, Alignment)

NonPowerOfTwoAlignAttribute(Span)

InconsistentBinding(Span)

TypeNotConstructible(Span)

TypeNotInferable(Span)

InitializationTypeMismatch

Fields

name: Span

expected: Box<str>

got: Box<str>

DeclMissingTypeAndInit(Span)

MissingAttribute(&'static str, Span)

InvalidAtomicPointer(Span)

InvalidAtomicOperandType(Span)

InvalidRayQueryPointer(Span)

Pointer(&'static str, Span)

NotPointer(Span)

NotReference(&'static str, Span)

InvalidAssignment

Fields

span: Span

ty: InvalidAssignmentType

ReservedKeyword(Span)

Redefinition

Redefinition of an identifier (used for both module-scope and local redefinitions).

Fields

previous: Span

Span of the identifier in the previous definition.

current: Span

Span of the identifier in the new definition.

RecursiveDeclaration

A declaration refers to itself directly.

Fields

ident: Span

The location of the name of the declaration.

usage: Span

The point at which it is used.

CyclicDeclaration

A declaration refers to itself indirectly, through one or more other definitions.

Fields

ident: Span

The location of the name of some declaration in the cycle.

path: Box<[(Span, Span)]>

The edges of the cycle of references.

Each (decl, reference) pair indicates that the declaration whose name is decl has an identifier at reference whose definition is the next declaration in the cycle. The last pair’s reference is the same identifier as ident, above.

InvalidSwitchValue

Fields

uint: bool

span: Span

CalledEntryPoint(Span)

WrongArgumentCount

Fields

span: Span

expected: Range<u32>

found: u32

FunctionReturnsVoid(Span)

InvalidWorkGroupUniformLoad(Span)

Internal(&'static str)

ExpectedConstExprConcreteIntegerScalar(Span)

ExpectedNonNegative(Span)

ExpectedPositiveArrayLength(Span)

MissingWorkgroupSize(Span)

ConstantEvaluatorError(Box<ConstantEvaluatorError>, Span)

AutoConversion(Box<AutoConversionError>)

AutoConversionLeafScalar(Box<AutoConversionLeafScalarError>)

ConcretizationFailed(Box<ConcretizationFailedError>)

ExceededLimitForNestedBraces

Fields

span: Span

limit: u8

PipelineConstantIDValue(Span)

NotBool(Span)

ConstAssertFailed(Span)

Implementations

impl<'a> Error<'a>

pub(crate) fn as_parse_error(&self, source: &'a str) -> ParseError