Module naga::proc

Module processing functionality.

Structs

InvalidBaseType

Layouter

Helper processor that derives the sizes of all types. It uses the default layout algorithm/table, described in https://github.com/gpuweb/gpuweb/issues/1393

Namer

This processor assigns names to all the things in a module that may need identifiers in a textual backend.

ResolveContext

TypeLayout

Alignment information for a type.

Enums

IndexableLength

The number of elements in an indexable type.

NameKey

ProcError

ResolveError

TypeResolution

Constants

POINTER_SPAN

Functions

ensure_block_returns

Ensure that the given block has return statements at the end of its control flow.

Type Definitions

Alignment

EntryPointIndex