Module naga::proc

Module processing functionality.

Structs

CallGraphBuilder

Interface

Layouter
Helper processor that derives the sizes of all types.

Namer
This processor assigns names to all the things in a module that may need identifiers in a textual backend.

ResolveContext

Typifier
Helper processor that derives the types of all expressions.

Validator

Enums

NameKey

ResolveError

ValidationError

Traits

Visitor

Functions

ensure_block_returns
Ensure that the given block has return statements at the end of its control flow.

Type Definitions

Alignment

CallGraph

EntryPointIndex