Module naga::proc

Module processing functionality.

Re-exports

pub use index::BoundsCheckPolicies;

pub use index::BoundsCheckPolicy;

pub use index::IndexableLength;

pub use index::IndexableLengthError;

Modules

index

Definitions for index bounds checking.

Structs

Alignment

A newtype struct where its only valid values are powers of 2

LayoutError

Layouter

Helper processor that derives the sizes of all types.

Namer

This processor assigns names to all the things in a module that may need identifiers in a textual backend.

ResolveContext

TypeLayout

Size and alignment information for a type.

Enums

LayoutErrorInner

NameKey

ResolveError

TypeResolution

The result of computing an expression’s type.

Constants

POINTER_SPAN

Functions

ensure_block_returns

Ensure that the given block has return statements at the end of its control flow.

Type Definitions

EntryPointIndex