naga::proc
Module constant_evaluator

Macros

__with_dollar_sign 🔒

__with_dollar_sign 🔒

__with_dollar_sign 🔒

__with_dollar_sign 🔒

component_wise_concrete_int 🔒

A convenience macro for using the same RHS for each ConcreteInt variant in a call to component_wise_concrete_int.

component_wise_float 🔒

A convenience macro for using the same RHS for each Float variant in a call to component_wise_float.

component_wise_scalar 🔒

A convenience macro for using the same RHS for each Scalar variant in a call to component_wise_scalar.

component_wise_signed 🔒

A convenience macro for using the same RHS for each Signed variant in a call to component_wise_signed.

gen_component_wise_extractor 🔒

with_dollar_sign 🔒

A macro that allows dollar signs ($) to be emitted by other macros. Useful for generating macro_rules! items that, in turn, emit their own macro_rules! items.

Structs

ConstantEvaluator

A context for evaluating constant expressions.

ExpressionKindTracker

FunctionLocalData 🔒

Enums

Behavior 🔒

ConcreteInt 🔒

A subset of Literals intended to be used for implementing numeric built-ins.

ConstantEvaluatorError

ExpressionKind

Float 🔒

A subset of Literals intended to be used for implementing numeric built-ins.

GlslRestrictions 🔒

Scalar 🔒

A subset of Literals intended to be used for implementing numeric built-ins.

Signed 🔒

A subset of Literals intended to be used for implementing numeric built-ins.

WgslRestrictions 🔒

Traits

TryFromAbstract 🔒

Trait for conversions of abstract values to concrete types.

Functions

component_wise_concrete_int 🔒

Attempts to evaluate multiple exprs as a combined ConcreteInt to pass to handler. If exprs are vectors of the same length, handler is called for each corresponding component of each vector.

component_wise_float 🔒

Attempts to evaluate multiple exprs as a combined Float to pass to handler. If exprs are vectors of the same length, handler is called for each corresponding component of each vector.

component_wise_scalar 🔒

Attempts to evaluate multiple exprs as a combined Scalar to pass to handler. If exprs are vectors of the same length, handler is called for each corresponding component of each vector.

component_wise_signed 🔒

Attempts to evaluate multiple exprs as a combined Signed to pass to handler. If exprs are vectors of the same length, handler is called for each corresponding component of each vector.

first_leading_bit 🔒

first_trailing_bit 🔒