naga::proc

Module index

Definitions for index bounds checking.

Structs

BoundsCheck 🔒

Items returned by the bounds_check_iter iterator.

BoundsCheckPolicies

Policies for injecting bounds checks during code generation.

Enums

BoundsCheckPolicy

How should code generated by Naga do bounds checks?

GuardedIndex

An index that may be statically known, or may need to be computed at runtime.

IndexableLength

The number of elements in an indexable type.

IndexableLengthError

Functions

access_needs_check

Determine whether index is statically known to be in bounds for base.

bounds_check_iter 🔒

Returns an iterator of accesses within the chain of Access and AccessIndex expressions starting from chain that may need to be bounds-checked at runtime.

find_checked_indexes

Build a set of expressions used as indices, to cache in temporary variables when emitted.

oob_local_types

Returns all the types which we need out-of-bounds locals for; that is, all of the types which the code might attempt to get an out-of-bounds pointer to, in which case we yield a pointer to the out-of-bounds local of the correct type.