Module naga::proc::index

Definitions for index bounds checking.

Structs

BoundsCheckPolicies

Policies for injecting bounds checks during code generation.

Enums

BoundsCheckPolicy

How should code generated by Naga do bounds checks?

GuardedIndex

An index that may be statically known, or may need to be computed at runtime.

IndexableLength

The number of elements in an indexable type.

IndexableLengthError

Functions

access_needs_check

Determine whether index is statically known to be in bounds for base.

find_checked_indexes

Build a set of expressions used as indices, to cache in temporary variables when emitted.