Struct naga::proc::ConstantEvaluator

pub struct ConstantEvaluator<'a> { /* private fields */ }

A context for evaluating constant expressions.

A ConstantEvaluator points at an expression arena to which it can append newly evaluated expressions: you pass try_eval_and_append whatever kind of Naga Expression you like, and if its value can be computed at compile time, try_eval_and_append appends an expression representing the computed value - a tree of Literal, Compose, ZeroValue, and Swizzle expressions - to the arena. See the try_eval_and_append method for details.

A ConstantEvaluator also holds whatever information we need to carry out that evaluation: types, other constants, and so on.

Implementations

impl<'a> ConstantEvaluator<'a>

pub fn for_wgsl_module(module: &'a mut Module) -> Self

Return a ConstantEvaluator that will add expressions to module’s constant expression arena.

Report errors according to WGSL’s rules for constant evaluation.

pub fn for_glsl_module(module: &'a mut Module) -> Self

Return a ConstantEvaluator that will add expressions to module’s constant expression arena.

Report errors according to GLSL’s rules for constant evaluation.

pub fn for_wgsl_function( module: &'a mut Module, expressions: &'a mut Arena<Expression>, expression_constness: &'a mut ExpressionConstnessTracker, emitter: &'a mut Emitter, block: &'a mut Block ) -> Self

Return a ConstantEvaluator that will add expressions to function’s expression arena.

Report errors according to WGSL’s rules for constant evaluation.

pub fn for_glsl_function( module: &'a mut Module, expressions: &'a mut Arena<Expression>, expression_constness: &'a mut ExpressionConstnessTracker, emitter: &'a mut Emitter, block: &'a mut Block ) -> Self

Return a ConstantEvaluator that will add expressions to function’s expression arena.

Report errors according to GLSL’s rules for constant evaluation.

pub fn try_eval_and_append( &mut self, expr: &Expression, span: Span ) -> Result<Handle<Expression>, ConstantEvaluatorError>

Try to evaluate expr at compile time.

The expr argument can be any sort of Naga Expression you like. If we can determine its value at compile time, we append an expression representing its value - a tree of Literal, Compose, ZeroValue, and Swizzle expressions - to the expression arena self contributes to.

If expr’s value cannot be determined at compile time, return a an error. If it’s acceptable to evaluate expr at runtime, this error can be ignored, and the caller can append expr to the arena itself.

We only consider expr itself, without recursing into its operands. Its operands must all have been produced by prior calls to try_eval_and_append, to ensure that they have already been reduced to an evaluated form if possible.

pub fn cast( &mut self, expr: Handle<Expression>, kind: ScalarKind, target_width: Bytes, span: Span ) -> Result<Handle<Expression>, ConstantEvaluatorError>

Convert the scalar components of expr to kind and target_width.

Treat span as the location of the resulting expression.