naga::proc::type_methods

Trait IntFloatLimits

pub trait IntFloatLimits<F>
where
    F: Float,
{
    // Required methods
    fn min_float() -> F;
    fn max_float() -> F;
}

Helper trait for providing the min and max values exactly representable by the integer type Self and floating point type F.

Required Methods

fn min_float() -> F

Returns the minimum value exactly representable by the integer type Self and floating point type F.

fn max_float() -> F

Returns the maximum value exactly representable by the integer type Self and floating point type F.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.